Type Definition tc37xpd::p00::Esr

pub type Esr = RegValueT<Esr_SPEC>;

Port 00 Emergency Stop Register resetvalue={:0x0}

Implementations

impl Esr

pub fn en0(self) -> RegisterField<0, 0x1, 1, 0, En0, Esr_SPEC, RW>

Emergency Stop Enable for Pin 15. This bit enables the emergency stop function for all GPIO lines. If the emergency stop condition is met and enabled the output selection is automatically switched from alternate output function to GPIO input function.

pub fn en1(self) -> RegisterField<1, 0x1, 1, 0, En1, Esr_SPEC, RW>

Emergency Stop Enable for Pin 15. This bit enables the emergency stop function for all GPIO lines. If the emergency stop condition is met and enabled the output selection is automatically switched from alternate output function to GPIO input function.

pub fn en2(self) -> RegisterField<2, 0x1, 1, 0, En2, Esr_SPEC, RW>

Emergency Stop Enable for Pin 15. This bit enables the emergency stop function for all GPIO lines. If the emergency stop condition is met and enabled the output selection is automatically switched from alternate output function to GPIO input function.

pub fn en3(self) -> RegisterField<3, 0x1, 1, 0, En3, Esr_SPEC, RW>

Emergency Stop Enable for Pin 15. This bit enables the emergency stop function for all GPIO lines. If the emergency stop condition is met and enabled the output selection is automatically switched from alternate output function to GPIO input function.

pub fn en4(self) -> RegisterField<4, 0x1, 1, 0, En4, Esr_SPEC, RW>

Emergency Stop Enable for Pin 15. This bit enables the emergency stop function for all GPIO lines. If the emergency stop condition is met and enabled the output selection is automatically switched from alternate output function to GPIO input function.

pub fn en5(self) -> RegisterField<5, 0x1, 1, 0, En5, Esr_SPEC, RW>

Emergency Stop Enable for Pin 15. This bit enables the emergency stop function for all GPIO lines. If the emergency stop condition is met and enabled the output selection is automatically switched from alternate output function to GPIO input function.

pub fn en6(self) -> RegisterField<6, 0x1, 1, 0, En6, Esr_SPEC, RW>

Emergency Stop Enable for Pin 15. This bit enables the emergency stop function for all GPIO lines. If the emergency stop condition is met and enabled the output selection is automatically switched from alternate output function to GPIO input function.

pub fn en7(self) -> RegisterField<7, 0x1, 1, 0, En7, Esr_SPEC, RW>

Emergency Stop Enable for Pin 15. This bit enables the emergency stop function for all GPIO lines. If the emergency stop condition is met and enabled the output selection is automatically switched from alternate output function to GPIO input function.

pub fn en8(self) -> RegisterField<8, 0x1, 1, 0, En8, Esr_SPEC, RW>

Emergency Stop Enable for Pin 15. This bit enables the emergency stop function for all GPIO lines. If the emergency stop condition is met and enabled the output selection is automatically switched from alternate output function to GPIO input function.

pub fn en9(self) -> RegisterField<9, 0x1, 1, 0, En9, Esr_SPEC, RW>

Emergency Stop Enable for Pin 15. This bit enables the emergency stop function for all GPIO lines. If the emergency stop condition is met and enabled the output selection is automatically switched from alternate output function to GPIO input function.

pub fn en10(self) -> RegisterField<10, 0x1, 1, 0, En10, Esr_SPEC, RW>

Emergency Stop Enable for Pin 15. This bit enables the emergency stop function for all GPIO lines. If the emergency stop condition is met and enabled the output selection is automatically switched from alternate output function to GPIO input function.

pub fn en11(self) -> RegisterField<11, 0x1, 1, 0, En11, Esr_SPEC, RW>

Emergency Stop Enable for Pin 15. This bit enables the emergency stop function for all GPIO lines. If the emergency stop condition is met and enabled the output selection is automatically switched from alternate output function to GPIO input function.

pub fn en12(self) -> RegisterField<12, 0x1, 1, 0, En12, Esr_SPEC, RW>

Emergency Stop Enable for Pin 15. This bit enables the emergency stop function for all GPIO lines. If the emergency stop condition is met and enabled the output selection is automatically switched from alternate output function to GPIO input function.

pub fn en13(self) -> RegisterField<13, 0x1, 1, 0, En13, Esr_SPEC, RW>

Emergency Stop Enable for Pin 15. This bit enables the emergency stop function for all GPIO lines. If the emergency stop condition is met and enabled the output selection is automatically switched from alternate output function to GPIO input function.

pub fn en14(self) -> RegisterField<14, 0x1, 1, 0, En14, Esr_SPEC, RW>

Emergency Stop Enable for Pin 15. This bit enables the emergency stop function for all GPIO lines. If the emergency stop condition is met and enabled the output selection is automatically switched from alternate output function to GPIO input function.

pub fn en15(self) -> RegisterField<15, 0x1, 1, 0, En15, Esr_SPEC, RW>

Emergency Stop Enable for Pin 15. This bit enables the emergency stop function for all GPIO lines. If the emergency stop condition is met and enabled the output selection is automatically switched from alternate output function to GPIO input function.