Type Definition tc37xpd::p00::In

pub type In = RegValueT<In_SPEC>;

Port 00 Input Register resetvalue={:0x0}

Implementations

impl In

pub fn p0(self) -> RegisterField<0, 0x1, 1, 0, P0, In_SPEC, R>

Input Bit 15. This bit indicates the level at the input pin Pn.x.

pub fn p1(self) -> RegisterField<1, 0x1, 1, 0, P1, In_SPEC, R>

Input Bit 15. This bit indicates the level at the input pin Pn.x.

pub fn p2(self) -> RegisterField<2, 0x1, 1, 0, P2, In_SPEC, R>

Input Bit 15. This bit indicates the level at the input pin Pn.x.

pub fn p3(self) -> RegisterField<3, 0x1, 1, 0, P3, In_SPEC, R>

Input Bit 15. This bit indicates the level at the input pin Pn.x.

pub fn p4(self) -> RegisterField<4, 0x1, 1, 0, P4, In_SPEC, R>

Input Bit 15. This bit indicates the level at the input pin Pn.x.

pub fn p5(self) -> RegisterField<5, 0x1, 1, 0, P5, In_SPEC, R>

Input Bit 15. This bit indicates the level at the input pin Pn.x.

pub fn p6(self) -> RegisterField<6, 0x1, 1, 0, P6, In_SPEC, R>

Input Bit 15. This bit indicates the level at the input pin Pn.x.

pub fn p7(self) -> RegisterField<7, 0x1, 1, 0, P7, In_SPEC, R>

Input Bit 15. This bit indicates the level at the input pin Pn.x.

pub fn p8(self) -> RegisterField<8, 0x1, 1, 0, P8, In_SPEC, R>

Input Bit 15. This bit indicates the level at the input pin Pn.x.

pub fn p9(self) -> RegisterField<9, 0x1, 1, 0, P9, In_SPEC, R>

Input Bit 15. This bit indicates the level at the input pin Pn.x.

pub fn p10(self) -> RegisterField<10, 0x1, 1, 0, P10, In_SPEC, R>

Input Bit 15. This bit indicates the level at the input pin Pn.x.

pub fn p11(self) -> RegisterField<11, 0x1, 1, 0, P11, In_SPEC, R>

Input Bit 15. This bit indicates the level at the input pin Pn.x.

pub fn p12(self) -> RegisterField<12, 0x1, 1, 0, P12, In_SPEC, R>

Input Bit 15. This bit indicates the level at the input pin Pn.x.

pub fn p13(self) -> RegisterField<13, 0x1, 1, 0, P13, In_SPEC, R>

Input Bit 15. This bit indicates the level at the input pin Pn.x.

pub fn p14(self) -> RegisterField<14, 0x1, 1, 0, P14, In_SPEC, R>

Input Bit 15. This bit indicates the level at the input pin Pn.x.

pub fn p15(self) -> RegisterField<15, 0x1, 1, 0, P15, In_SPEC, R>

Input Bit 15. This bit indicates the level at the input pin Pn.x.