Type Definition tc37xpd::p00::Omcr

pub type Omcr = RegValueT<Omcr_SPEC>;

Port 00 Output Modification Clear Register resetvalue={:0x0}

Implementations

impl Omcr

pub fn pcl0(self) -> RegisterField<16, 0x1, 1, 0, Pcl0, Omcr_SPEC, W>

Clear Bit 15. Setting this bit will clear the corresponding bit in the port output register Pn OUT. Read as 0.

pub fn pcl1(self) -> RegisterField<17, 0x1, 1, 0, Pcl1, Omcr_SPEC, W>

Clear Bit 15. Setting this bit will clear the corresponding bit in the port output register Pn OUT. Read as 0.

pub fn pcl2(self) -> RegisterField<18, 0x1, 1, 0, Pcl2, Omcr_SPEC, W>

Clear Bit 15. Setting this bit will clear the corresponding bit in the port output register Pn OUT. Read as 0.

pub fn pcl3(self) -> RegisterField<19, 0x1, 1, 0, Pcl3, Omcr_SPEC, W>

Clear Bit 15. Setting this bit will clear the corresponding bit in the port output register Pn OUT. Read as 0.

pub fn pcl4(self) -> RegisterField<20, 0x1, 1, 0, Pcl4, Omcr_SPEC, W>

Clear Bit 15. Setting this bit will clear the corresponding bit in the port output register Pn OUT. Read as 0.

pub fn pcl5(self) -> RegisterField<21, 0x1, 1, 0, Pcl5, Omcr_SPEC, W>

Clear Bit 15. Setting this bit will clear the corresponding bit in the port output register Pn OUT. Read as 0.

pub fn pcl6(self) -> RegisterField<22, 0x1, 1, 0, Pcl6, Omcr_SPEC, W>

Clear Bit 15. Setting this bit will clear the corresponding bit in the port output register Pn OUT. Read as 0.

pub fn pcl7(self) -> RegisterField<23, 0x1, 1, 0, Pcl7, Omcr_SPEC, W>

Clear Bit 15. Setting this bit will clear the corresponding bit in the port output register Pn OUT. Read as 0.

pub fn pcl8(self) -> RegisterField<24, 0x1, 1, 0, Pcl8, Omcr_SPEC, W>

Clear Bit 15. Setting this bit will clear the corresponding bit in the port output register Pn OUT. Read as 0.

pub fn pcl9(self) -> RegisterField<25, 0x1, 1, 0, Pcl9, Omcr_SPEC, W>

Clear Bit 15. Setting this bit will clear the corresponding bit in the port output register Pn OUT. Read as 0.

pub fn pcl10(self) -> RegisterField<26, 0x1, 1, 0, Pcl10, Omcr_SPEC, W>

Clear Bit 15. Setting this bit will clear the corresponding bit in the port output register Pn OUT. Read as 0.

pub fn pcl11(self) -> RegisterField<27, 0x1, 1, 0, Pcl11, Omcr_SPEC, W>

Clear Bit 15. Setting this bit will clear the corresponding bit in the port output register Pn OUT. Read as 0.

pub fn pcl12(self) -> RegisterField<28, 0x1, 1, 0, Pcl12, Omcr_SPEC, W>

Clear Bit 15. Setting this bit will clear the corresponding bit in the port output register Pn OUT. Read as 0.

pub fn pcl13(self) -> RegisterField<29, 0x1, 1, 0, Pcl13, Omcr_SPEC, W>

Clear Bit 15. Setting this bit will clear the corresponding bit in the port output register Pn OUT. Read as 0.

pub fn pcl14(self) -> RegisterField<30, 0x1, 1, 0, Pcl14, Omcr_SPEC, W>

Clear Bit 15. Setting this bit will clear the corresponding bit in the port output register Pn OUT. Read as 0.

pub fn pcl15(self) -> RegisterField<31, 0x1, 1, 0, Pcl15, Omcr_SPEC, W>

Clear Bit 15. Setting this bit will clear the corresponding bit in the port output register Pn OUT. Read as 0.