Type Definition tc37xpd::p00::Omr

pub type Omr = RegValueT<Omr_SPEC>;

Port 00 Output Modification Register resetvalue={:0x0}

Implementations

impl Omr

pub fn ps0(self) -> RegisterField<0, 0x1, 1, 0, Ps0, Omr_SPEC, W>

Set Bit 15. Setting this bit will set or toggle the corresponding bit in the port output register Pn OUT. Read as 0. The function of this bit is shown in Table .

pub fn ps1(self) -> RegisterField<1, 0x1, 1, 0, Ps1, Omr_SPEC, W>

Set Bit 15. Setting this bit will set or toggle the corresponding bit in the port output register Pn OUT. Read as 0. The function of this bit is shown in Table .

pub fn ps2(self) -> RegisterField<2, 0x1, 1, 0, Ps2, Omr_SPEC, W>

Set Bit 15. Setting this bit will set or toggle the corresponding bit in the port output register Pn OUT. Read as 0. The function of this bit is shown in Table .

pub fn ps3(self) -> RegisterField<3, 0x1, 1, 0, Ps3, Omr_SPEC, W>

Set Bit 15. Setting this bit will set or toggle the corresponding bit in the port output register Pn OUT. Read as 0. The function of this bit is shown in Table .

pub fn ps4(self) -> RegisterField<4, 0x1, 1, 0, Ps4, Omr_SPEC, W>

Set Bit 15. Setting this bit will set or toggle the corresponding bit in the port output register Pn OUT. Read as 0. The function of this bit is shown in Table .

pub fn ps5(self) -> RegisterField<5, 0x1, 1, 0, Ps5, Omr_SPEC, W>

Set Bit 15. Setting this bit will set or toggle the corresponding bit in the port output register Pn OUT. Read as 0. The function of this bit is shown in Table .

pub fn ps6(self) -> RegisterField<6, 0x1, 1, 0, Ps6, Omr_SPEC, W>

Set Bit 15. Setting this bit will set or toggle the corresponding bit in the port output register Pn OUT. Read as 0. The function of this bit is shown in Table .

pub fn ps7(self) -> RegisterField<7, 0x1, 1, 0, Ps7, Omr_SPEC, W>

Set Bit 15. Setting this bit will set or toggle the corresponding bit in the port output register Pn OUT. Read as 0. The function of this bit is shown in Table .

pub fn ps8(self) -> RegisterField<8, 0x1, 1, 0, Ps8, Omr_SPEC, W>

Set Bit 15. Setting this bit will set or toggle the corresponding bit in the port output register Pn OUT. Read as 0. The function of this bit is shown in Table .

pub fn ps9(self) -> RegisterField<9, 0x1, 1, 0, Ps9, Omr_SPEC, W>

Set Bit 15. Setting this bit will set or toggle the corresponding bit in the port output register Pn OUT. Read as 0. The function of this bit is shown in Table .

pub fn ps10(self) -> RegisterField<10, 0x1, 1, 0, Ps10, Omr_SPEC, W>

Set Bit 15. Setting this bit will set or toggle the corresponding bit in the port output register Pn OUT. Read as 0. The function of this bit is shown in Table .

pub fn ps11(self) -> RegisterField<11, 0x1, 1, 0, Ps11, Omr_SPEC, W>

Set Bit 15. Setting this bit will set or toggle the corresponding bit in the port output register Pn OUT. Read as 0. The function of this bit is shown in Table .

pub fn ps12(self) -> RegisterField<12, 0x1, 1, 0, Ps12, Omr_SPEC, W>

Set Bit 15. Setting this bit will set or toggle the corresponding bit in the port output register Pn OUT. Read as 0. The function of this bit is shown in Table .

pub fn ps13(self) -> RegisterField<13, 0x1, 1, 0, Ps13, Omr_SPEC, W>

Set Bit 15. Setting this bit will set or toggle the corresponding bit in the port output register Pn OUT. Read as 0. The function of this bit is shown in Table .

pub fn ps14(self) -> RegisterField<14, 0x1, 1, 0, Ps14, Omr_SPEC, W>

Set Bit 15. Setting this bit will set or toggle the corresponding bit in the port output register Pn OUT. Read as 0. The function of this bit is shown in Table .

pub fn ps15(self) -> RegisterField<15, 0x1, 1, 0, Ps15, Omr_SPEC, W>

Set Bit 15. Setting this bit will set or toggle the corresponding bit in the port output register Pn OUT. Read as 0. The function of this bit is shown in Table .

pub fn pcl0(self) -> RegisterField<16, 0x1, 1, 0, Pcl0, Omr_SPEC, W>

Clear Bit 15. Setting this bit will clear or toggle the corresponding bit in the port output register Pn OUT. Read as 0. The function of this bit is shown in Table .

pub fn pcl1(self) -> RegisterField<17, 0x1, 1, 0, Pcl1, Omr_SPEC, W>

Clear Bit 15. Setting this bit will clear or toggle the corresponding bit in the port output register Pn OUT. Read as 0. The function of this bit is shown in Table .

pub fn pcl2(self) -> RegisterField<18, 0x1, 1, 0, Pcl2, Omr_SPEC, W>

Clear Bit 15. Setting this bit will clear or toggle the corresponding bit in the port output register Pn OUT. Read as 0. The function of this bit is shown in Table .

pub fn pcl3(self) -> RegisterField<19, 0x1, 1, 0, Pcl3, Omr_SPEC, W>

Clear Bit 15. Setting this bit will clear or toggle the corresponding bit in the port output register Pn OUT. Read as 0. The function of this bit is shown in Table .

pub fn pcl4(self) -> RegisterField<20, 0x1, 1, 0, Pcl4, Omr_SPEC, W>

Clear Bit 15. Setting this bit will clear or toggle the corresponding bit in the port output register Pn OUT. Read as 0. The function of this bit is shown in Table .

pub fn pcl5(self) -> RegisterField<21, 0x1, 1, 0, Pcl5, Omr_SPEC, W>

Clear Bit 15. Setting this bit will clear or toggle the corresponding bit in the port output register Pn OUT. Read as 0. The function of this bit is shown in Table .

pub fn pcl6(self) -> RegisterField<22, 0x1, 1, 0, Pcl6, Omr_SPEC, W>

Clear Bit 15. Setting this bit will clear or toggle the corresponding bit in the port output register Pn OUT. Read as 0. The function of this bit is shown in Table .

pub fn pcl7(self) -> RegisterField<23, 0x1, 1, 0, Pcl7, Omr_SPEC, W>

Clear Bit 15. Setting this bit will clear or toggle the corresponding bit in the port output register Pn OUT. Read as 0. The function of this bit is shown in Table .

pub fn pcl8(self) -> RegisterField<24, 0x1, 1, 0, Pcl8, Omr_SPEC, W>

Clear Bit 15. Setting this bit will clear or toggle the corresponding bit in the port output register Pn OUT. Read as 0. The function of this bit is shown in Table .

pub fn pcl9(self) -> RegisterField<25, 0x1, 1, 0, Pcl9, Omr_SPEC, W>

Clear Bit 15. Setting this bit will clear or toggle the corresponding bit in the port output register Pn OUT. Read as 0. The function of this bit is shown in Table .

pub fn pcl10(self) -> RegisterField<26, 0x1, 1, 0, Pcl10, Omr_SPEC, W>

Clear Bit 15. Setting this bit will clear or toggle the corresponding bit in the port output register Pn OUT. Read as 0. The function of this bit is shown in Table .

pub fn pcl11(self) -> RegisterField<27, 0x1, 1, 0, Pcl11, Omr_SPEC, W>

Clear Bit 15. Setting this bit will clear or toggle the corresponding bit in the port output register Pn OUT. Read as 0. The function of this bit is shown in Table .

pub fn pcl12(self) -> RegisterField<28, 0x1, 1, 0, Pcl12, Omr_SPEC, W>

Clear Bit 15. Setting this bit will clear or toggle the corresponding bit in the port output register Pn OUT. Read as 0. The function of this bit is shown in Table .

pub fn pcl13(self) -> RegisterField<29, 0x1, 1, 0, Pcl13, Omr_SPEC, W>

Clear Bit 15. Setting this bit will clear or toggle the corresponding bit in the port output register Pn OUT. Read as 0. The function of this bit is shown in Table .

pub fn pcl14(self) -> RegisterField<30, 0x1, 1, 0, Pcl14, Omr_SPEC, W>

Clear Bit 15. Setting this bit will clear or toggle the corresponding bit in the port output register Pn OUT. Read as 0. The function of this bit is shown in Table .

pub fn pcl15(self) -> RegisterField<31, 0x1, 1, 0, Pcl15, Omr_SPEC, W>

Clear Bit 15. Setting this bit will clear or toggle the corresponding bit in the port output register Pn OUT. Read as 0. The function of this bit is shown in Table .