Type Definition tc37xpd::p00::Out

pub type Out = RegValueT<Out_SPEC>;

Port 00 Output Register resetvalue={:0x0}

Implementations

impl Out

pub fn p0(self) -> RegisterField<0, 0x1, 1, 0, P0, Out_SPEC, RW>

Output Bit 15. This bit determines the level at the output pin Pn.x if the output is selected as GPIO output. Pn.x can also be set or cleared by control bits of the Pn OMSR Pn OMCR or Pn OMR registers.

pub fn p1(self) -> RegisterField<1, 0x1, 1, 0, P1, Out_SPEC, RW>

Output Bit 15. This bit determines the level at the output pin Pn.x if the output is selected as GPIO output. Pn.x can also be set or cleared by control bits of the Pn OMSR Pn OMCR or Pn OMR registers.

pub fn p2(self) -> RegisterField<2, 0x1, 1, 0, P2, Out_SPEC, RW>

Output Bit 15. This bit determines the level at the output pin Pn.x if the output is selected as GPIO output. Pn.x can also be set or cleared by control bits of the Pn OMSR Pn OMCR or Pn OMR registers.

pub fn p3(self) -> RegisterField<3, 0x1, 1, 0, P3, Out_SPEC, RW>

Output Bit 15. This bit determines the level at the output pin Pn.x if the output is selected as GPIO output. Pn.x can also be set or cleared by control bits of the Pn OMSR Pn OMCR or Pn OMR registers.

pub fn p4(self) -> RegisterField<4, 0x1, 1, 0, P4, Out_SPEC, RW>

Output Bit 15. This bit determines the level at the output pin Pn.x if the output is selected as GPIO output. Pn.x can also be set or cleared by control bits of the Pn OMSR Pn OMCR or Pn OMR registers.

pub fn p5(self) -> RegisterField<5, 0x1, 1, 0, P5, Out_SPEC, RW>

Output Bit 15. This bit determines the level at the output pin Pn.x if the output is selected as GPIO output. Pn.x can also be set or cleared by control bits of the Pn OMSR Pn OMCR or Pn OMR registers.

pub fn p6(self) -> RegisterField<6, 0x1, 1, 0, P6, Out_SPEC, RW>

Output Bit 15. This bit determines the level at the output pin Pn.x if the output is selected as GPIO output. Pn.x can also be set or cleared by control bits of the Pn OMSR Pn OMCR or Pn OMR registers.

pub fn p7(self) -> RegisterField<7, 0x1, 1, 0, P7, Out_SPEC, RW>

Output Bit 15. This bit determines the level at the output pin Pn.x if the output is selected as GPIO output. Pn.x can also be set or cleared by control bits of the Pn OMSR Pn OMCR or Pn OMR registers.

pub fn p8(self) -> RegisterField<8, 0x1, 1, 0, P8, Out_SPEC, RW>

Output Bit 15. This bit determines the level at the output pin Pn.x if the output is selected as GPIO output. Pn.x can also be set or cleared by control bits of the Pn OMSR Pn OMCR or Pn OMR registers.

pub fn p9(self) -> RegisterField<9, 0x1, 1, 0, P9, Out_SPEC, RW>

Output Bit 15. This bit determines the level at the output pin Pn.x if the output is selected as GPIO output. Pn.x can also be set or cleared by control bits of the Pn OMSR Pn OMCR or Pn OMR registers.

pub fn p10(self) -> RegisterField<10, 0x1, 1, 0, P10, Out_SPEC, RW>

Output Bit 15. This bit determines the level at the output pin Pn.x if the output is selected as GPIO output. Pn.x can also be set or cleared by control bits of the Pn OMSR Pn OMCR or Pn OMR registers.

pub fn p11(self) -> RegisterField<11, 0x1, 1, 0, P11, Out_SPEC, RW>

Output Bit 15. This bit determines the level at the output pin Pn.x if the output is selected as GPIO output. Pn.x can also be set or cleared by control bits of the Pn OMSR Pn OMCR or Pn OMR registers.

pub fn p12(self) -> RegisterField<12, 0x1, 1, 0, P12, Out_SPEC, RW>

Output Bit 15. This bit determines the level at the output pin Pn.x if the output is selected as GPIO output. Pn.x can also be set or cleared by control bits of the Pn OMSR Pn OMCR or Pn OMR registers.

pub fn p13(self) -> RegisterField<13, 0x1, 1, 0, P13, Out_SPEC, RW>

Output Bit 15. This bit determines the level at the output pin Pn.x if the output is selected as GPIO output. Pn.x can also be set or cleared by control bits of the Pn OMSR Pn OMCR or Pn OMR registers.

pub fn p14(self) -> RegisterField<14, 0x1, 1, 0, P14, Out_SPEC, RW>

Output Bit 15. This bit determines the level at the output pin Pn.x if the output is selected as GPIO output. Pn.x can also be set or cleared by control bits of the Pn OMSR Pn OMCR or Pn OMR registers.

pub fn p15(self) -> RegisterField<15, 0x1, 1, 0, P15, Out_SPEC, RW>

Output Bit 15. This bit determines the level at the output pin Pn.x if the output is selected as GPIO output. Pn.x can also be set or cleared by control bits of the Pn OMSR Pn OMCR or Pn OMR registers.