Type Definition tc37xpd::p00::Pdisc

pub type Pdisc = RegValueT<Pdisc_SPEC>;

Port 00 Pin Function Decision Control Register resetvalue={:0x0,:0x0,After SSW execution:0x0,After SSW execution:0x0}

Implementations

impl Pdisc

pub fn pdis0(self) -> RegisterField<0, 0x1, 1, 0, Pdis0, Pdisc_SPEC, RW>

Pin Function Decision Control for Pin 15. This bit selects the function of the port pad.

pub fn pdis1(self) -> RegisterField<1, 0x1, 1, 0, Pdis1, Pdisc_SPEC, RW>

Pin Function Decision Control for Pin 15. This bit selects the function of the port pad.

pub fn pdis2(self) -> RegisterField<2, 0x1, 1, 0, Pdis2, Pdisc_SPEC, RW>

Pin Function Decision Control for Pin 15. This bit selects the function of the port pad.

pub fn pdis3(self) -> RegisterField<3, 0x1, 1, 0, Pdis3, Pdisc_SPEC, RW>

Pin Function Decision Control for Pin 15. This bit selects the function of the port pad.

pub fn pdis4(self) -> RegisterField<4, 0x1, 1, 0, Pdis4, Pdisc_SPEC, RW>

Pin Function Decision Control for Pin 15. This bit selects the function of the port pad.

pub fn pdis5(self) -> RegisterField<5, 0x1, 1, 0, Pdis5, Pdisc_SPEC, RW>

Pin Function Decision Control for Pin 15. This bit selects the function of the port pad.

pub fn pdis6(self) -> RegisterField<6, 0x1, 1, 0, Pdis6, Pdisc_SPEC, RW>

Pin Function Decision Control for Pin 15. This bit selects the function of the port pad.

pub fn pdis7(self) -> RegisterField<7, 0x1, 1, 0, Pdis7, Pdisc_SPEC, RW>

Pin Function Decision Control for Pin 15. This bit selects the function of the port pad.

pub fn pdis8(self) -> RegisterField<8, 0x1, 1, 0, Pdis8, Pdisc_SPEC, RW>

Pin Function Decision Control for Pin 15. This bit selects the function of the port pad.

pub fn pdis9(self) -> RegisterField<9, 0x1, 1, 0, Pdis9, Pdisc_SPEC, RW>

Pin Function Decision Control for Pin 15. This bit selects the function of the port pad.

pub fn pdis10(self) -> RegisterField<10, 0x1, 1, 0, Pdis10, Pdisc_SPEC, RW>

Pin Function Decision Control for Pin 15. This bit selects the function of the port pad.

pub fn pdis11(self) -> RegisterField<11, 0x1, 1, 0, Pdis11, Pdisc_SPEC, RW>

Pin Function Decision Control for Pin 15. This bit selects the function of the port pad.

pub fn pdis12(self) -> RegisterField<12, 0x1, 1, 0, Pdis12, Pdisc_SPEC, RW>

Pin Function Decision Control for Pin 15. This bit selects the function of the port pad.

pub fn pdis13(self) -> RegisterField<13, 0x1, 1, 0, Pdis13, Pdisc_SPEC, RW>

Pin Function Decision Control for Pin 15. This bit selects the function of the port pad.

pub fn pdis14(self) -> RegisterField<14, 0x1, 1, 0, Pdis14, Pdisc_SPEC, RW>

Pin Function Decision Control for Pin 15. This bit selects the function of the port pad.

pub fn pdis15(self) -> RegisterField<15, 0x1, 1, 0, Pdis15, Pdisc_SPEC, RW>

Pin Function Decision Control for Pin 15. This bit selects the function of the port pad.