Type Definition tc37xpd::scu::Accen00

pub type Accen00 = RegValueT<Accen00_SPEC>;

Access Enable Register 00 resetvalue={Application Reset:0x0FFFFFFFF}

Implementations

impl Accen00

pub fn en0(self) -> RegisterField<0, 0x1, 1, 0, En0, Accen00_SPEC, RW>

Access Enable for Master TAG ID 31 EN31. This bit enables write access to the SCU kernel addresses for transactions with the Master TAG ID n

pub fn en1(self) -> RegisterField<1, 0x1, 1, 0, En1, Accen00_SPEC, RW>

Access Enable for Master TAG ID 31 EN31. This bit enables write access to the SCU kernel addresses for transactions with the Master TAG ID n

pub fn en2(self) -> RegisterField<2, 0x1, 1, 0, En2, Accen00_SPEC, RW>

Access Enable for Master TAG ID 31 EN31. This bit enables write access to the SCU kernel addresses for transactions with the Master TAG ID n

pub fn en3(self) -> RegisterField<3, 0x1, 1, 0, En3, Accen00_SPEC, RW>

Access Enable for Master TAG ID 31 EN31. This bit enables write access to the SCU kernel addresses for transactions with the Master TAG ID n

pub fn en4(self) -> RegisterField<4, 0x1, 1, 0, En4, Accen00_SPEC, RW>

Access Enable for Master TAG ID 31 EN31. This bit enables write access to the SCU kernel addresses for transactions with the Master TAG ID n

pub fn en5(self) -> RegisterField<5, 0x1, 1, 0, En5, Accen00_SPEC, RW>

Access Enable for Master TAG ID 31 EN31. This bit enables write access to the SCU kernel addresses for transactions with the Master TAG ID n

pub fn en6(self) -> RegisterField<6, 0x1, 1, 0, En6, Accen00_SPEC, RW>

Access Enable for Master TAG ID 31 EN31. This bit enables write access to the SCU kernel addresses for transactions with the Master TAG ID n

pub fn en7(self) -> RegisterField<7, 0x1, 1, 0, En7, Accen00_SPEC, RW>

Access Enable for Master TAG ID 31 EN31. This bit enables write access to the SCU kernel addresses for transactions with the Master TAG ID n

pub fn en8(self) -> RegisterField<8, 0x1, 1, 0, En8, Accen00_SPEC, RW>

Access Enable for Master TAG ID 31 EN31. This bit enables write access to the SCU kernel addresses for transactions with the Master TAG ID n

pub fn en9(self) -> RegisterField<9, 0x1, 1, 0, En9, Accen00_SPEC, RW>

Access Enable for Master TAG ID 31 EN31. This bit enables write access to the SCU kernel addresses for transactions with the Master TAG ID n

pub fn en10(self) -> RegisterField<10, 0x1, 1, 0, En10, Accen00_SPEC, RW>

Access Enable for Master TAG ID 31 EN31. This bit enables write access to the SCU kernel addresses for transactions with the Master TAG ID n

pub fn en11(self) -> RegisterField<11, 0x1, 1, 0, En11, Accen00_SPEC, RW>

Access Enable for Master TAG ID 31 EN31. This bit enables write access to the SCU kernel addresses for transactions with the Master TAG ID n

pub fn en12(self) -> RegisterField<12, 0x1, 1, 0, En12, Accen00_SPEC, RW>

Access Enable for Master TAG ID 31 EN31. This bit enables write access to the SCU kernel addresses for transactions with the Master TAG ID n

pub fn en13(self) -> RegisterField<13, 0x1, 1, 0, En13, Accen00_SPEC, RW>

Access Enable for Master TAG ID 31 EN31. This bit enables write access to the SCU kernel addresses for transactions with the Master TAG ID n

pub fn en14(self) -> RegisterField<14, 0x1, 1, 0, En14, Accen00_SPEC, RW>

Access Enable for Master TAG ID 31 EN31. This bit enables write access to the SCU kernel addresses for transactions with the Master TAG ID n

pub fn en15(self) -> RegisterField<15, 0x1, 1, 0, En15, Accen00_SPEC, RW>

Access Enable for Master TAG ID 31 EN31. This bit enables write access to the SCU kernel addresses for transactions with the Master TAG ID n

pub fn en16(self) -> RegisterField<16, 0x1, 1, 0, En16, Accen00_SPEC, RW>

Access Enable for Master TAG ID 31 EN31. This bit enables write access to the SCU kernel addresses for transactions with the Master TAG ID n

pub fn en17(self) -> RegisterField<17, 0x1, 1, 0, En17, Accen00_SPEC, RW>

Access Enable for Master TAG ID 31 EN31. This bit enables write access to the SCU kernel addresses for transactions with the Master TAG ID n

pub fn en18(self) -> RegisterField<18, 0x1, 1, 0, En18, Accen00_SPEC, RW>

Access Enable for Master TAG ID 31 EN31. This bit enables write access to the SCU kernel addresses for transactions with the Master TAG ID n

pub fn en19(self) -> RegisterField<19, 0x1, 1, 0, En19, Accen00_SPEC, RW>

Access Enable for Master TAG ID 31 EN31. This bit enables write access to the SCU kernel addresses for transactions with the Master TAG ID n

pub fn en20(self) -> RegisterField<20, 0x1, 1, 0, En20, Accen00_SPEC, RW>

Access Enable for Master TAG ID 31 EN31. This bit enables write access to the SCU kernel addresses for transactions with the Master TAG ID n

pub fn en21(self) -> RegisterField<21, 0x1, 1, 0, En21, Accen00_SPEC, RW>

Access Enable for Master TAG ID 31 EN31. This bit enables write access to the SCU kernel addresses for transactions with the Master TAG ID n

pub fn en22(self) -> RegisterField<22, 0x1, 1, 0, En22, Accen00_SPEC, RW>

Access Enable for Master TAG ID 31 EN31. This bit enables write access to the SCU kernel addresses for transactions with the Master TAG ID n

pub fn en23(self) -> RegisterField<23, 0x1, 1, 0, En23, Accen00_SPEC, RW>

Access Enable for Master TAG ID 31 EN31. This bit enables write access to the SCU kernel addresses for transactions with the Master TAG ID n

pub fn en24(self) -> RegisterField<24, 0x1, 1, 0, En24, Accen00_SPEC, RW>

Access Enable for Master TAG ID 31 EN31. This bit enables write access to the SCU kernel addresses for transactions with the Master TAG ID n

pub fn en25(self) -> RegisterField<25, 0x1, 1, 0, En25, Accen00_SPEC, RW>

Access Enable for Master TAG ID 31 EN31. This bit enables write access to the SCU kernel addresses for transactions with the Master TAG ID n

pub fn en26(self) -> RegisterField<26, 0x1, 1, 0, En26, Accen00_SPEC, RW>

Access Enable for Master TAG ID 31 EN31. This bit enables write access to the SCU kernel addresses for transactions with the Master TAG ID n

pub fn en27(self) -> RegisterField<27, 0x1, 1, 0, En27, Accen00_SPEC, RW>

Access Enable for Master TAG ID 31 EN31. This bit enables write access to the SCU kernel addresses for transactions with the Master TAG ID n

pub fn en28(self) -> RegisterField<28, 0x1, 1, 0, En28, Accen00_SPEC, RW>

Access Enable for Master TAG ID 31 EN31. This bit enables write access to the SCU kernel addresses for transactions with the Master TAG ID n

pub fn en29(self) -> RegisterField<29, 0x1, 1, 0, En29, Accen00_SPEC, RW>

Access Enable for Master TAG ID 31 EN31. This bit enables write access to the SCU kernel addresses for transactions with the Master TAG ID n

pub fn en30(self) -> RegisterField<30, 0x1, 1, 0, En30, Accen00_SPEC, RW>

Access Enable for Master TAG ID 31 EN31. This bit enables write access to the SCU kernel addresses for transactions with the Master TAG ID n

pub fn en31(self) -> RegisterField<31, 0x1, 1, 0, En31, Accen00_SPEC, RW>

Access Enable for Master TAG ID 31 EN31. This bit enables write access to the SCU kernel addresses for transactions with the Master TAG ID n