Type Definition tc37xpd::scu::Arstdis

pub type Arstdis = RegValueT<Arstdis_SPEC>;

Application Reset Disable Register resetvalue={PowerOn Reset:0x0}

Implementations

impl Arstdis

pub fn stm0dis(self) -> RegisterField<0, 0x1, 1, 0, Stm0Dis, Arstdis_SPEC, RW>

STM0 Disable Reset STM0DIS. This bit field defines if an Application Reset leads to an reset for the STM0.

pub fn stm1dis(self) -> RegisterField<1, 0x1, 1, 0, Stm1Dis, Arstdis_SPEC, RW>

STM1 Disable Reset STM1DIS. This bit field defines if an Application Reset leads to an reset for the STM1.

pub fn stm2dis(self) -> RegisterField<2, 0x1, 1, 0, Stm2Dis, Arstdis_SPEC, RW>

STM2 Disable Reset STM2DIS. This bit field defines if an Application Reset leads to an reset for the STM2.