Type Definition tc37xpd::scu::Ccucon0

pub type Ccucon0 = RegValueT<Ccucon0_SPEC>;

CCU Clock Control Register 0 resetvalue={After SSW execution:0x5120112,System Reset:0x3130113,System Reset:0x35120112}

Implementations

impl Ccucon0

pub fn stmdiv(self) -> RegisterField<0, 0xf, 1, 0, Stmdiv, Ccucon0_SPEC, RW>

STM Divider Reload Value STMDIV. The resulting STM frequency is configured to f STM xa0 xa0 f source0 xa0 xa0 STMDIV for the allowed configurations. For STMDIV xa0 xa0 0000 B the clock is shut off. f source0 can be configured either to f PLL0 CLKSEL xa0 xa0 01 B or f BACK CLKSEL xa0 xa0 00 B

pub fn gtmdiv(self) -> RegisterField<4, 0xf, 1, 0, Gtmdiv, Ccucon0_SPEC, RW>

GTM Divider Reload Value GTMDIV. The resulting GTM frequency is configured to f GTM f SOURCEGTM GTMDIV for the allowed configurations. For GTMDIV 0000 B the clock is shut off.

pub fn sridiv(self) -> RegisterField<8, 0xf, 1, 0, Sridiv, Ccucon0_SPEC, RW>

SRI Divider Reload Value SRIDIV. The resulting SRI frequency is configured to f SRI f source0 SRIDIV for the allowed configurations. f source0 could be configured either to f PLL0 CLKSEL 01 B or f BACK CLKSEL 00 B

pub fn lpdiv(self) -> RegisterField<12, 0x7, 1, 0, Lpdiv, Ccucon0_SPEC, RW>

Low Power Divider Reload Value LPDIV. The selected divider is valid for all clocks derived from f XXX with XXX SPB SRI BBB FSI GETH GTM ADAS.

pub fn spbdiv(self) -> RegisterField<16, 0xf, 1, 0, Spbdiv, Ccucon0_SPEC, RW>

SPB Divider Reload Value SPBDIV. f source0 could be configured either to f PLL CLKSEL 01 B or f BACK CLKSEL 00 B

pub fn bbbdiv(self) -> RegisterField<20, 0xf, 1, 0, Bbbdiv, Ccucon0_SPEC, RW>

BBB Divider Reload Value BBBDIV. The resulting BBB frequency is configured to f BBB 160 160 f source0 160 160 BBBDIV for all allowed configurations. For BBBDIV 160 160 0000 B the clock is shut off. f source0 could be configured either to f PLL0 CLKSEL 160 160 01 B or f BACK CLKSEL 160 160 00 B

pub fn fsidiv(self) -> RegisterField<24, 0x3, 1, 0, Fsidiv, Ccucon0_SPEC, RW>

FSI Divider Reload Value FSIDIV

pub fn fsi2div(self) -> RegisterField<26, 0x3, 1, 0, Fsi2Div, Ccucon0_SPEC, RW>

FSI2 Divider Reload Value FSI2DIV

pub fn clksel(self) -> RegisterField<28, 0x3, 1, 0, Clksel, Ccucon0_SPEC, RW>

Clock Selection for Source CLKSEL. This bit field defines the clock source that is used for the clock generation of f sourcex .

pub fn up(self) -> RegisterField<30, 0x1, 1, 0, Up, Ccucon0_SPEC, W>

Update Request UP. Setting this bit will request an update for CCUCON0 and CCUCON5. Only one UP bit must be set for either CCUCON0 or CCUCON5. This bit always reads as zero.

pub fn lck(self) -> RegisterField<31, 0x1, 1, 0, Lck, Ccucon0_SPEC, R>

Lock Status LCK. This bit indicates if the register can be updated with a new value or if the register is locked and a write action from the bus side has no effect. The lock bit is set when an update of CCUCON0 5 has been requested and released when the update is complete.