Type Definition tc37xpd::scu::Ccucon1

pub type Ccucon1 = RegValueT<Ccucon1_SPEC>;

CCU Clock Control Register 1 resetvalue={System Reset:0x10100302}

Implementations

impl Ccucon1

pub fn mcandiv(self) -> RegisterField<0, 0xf, 1, 0, Mcandiv, Ccucon1_SPEC, RW>

MCAN Divider Reload Value MCANDIV. The resulting MCAN frequency is configured to f MCANI 160 160 f source1 160 160 MCANDIV for the allowed configurations. For MCANDIV 160 160 0000 B the clock is shut off.

pub fn clkselmcan( self ) -> RegisterField<4, 0x3, 1, 0, Clkselmcan, Ccucon1_SPEC, RW>

Clock Selection for MCAN CLKSELMCAN. This bit field defines the clock source that is used for the clock generation of f SOURCEMCAN . For switching between two non zero configurations the following sequence has to be applied First step is to switch to 00 B . Second step is to switch to the new target configuration.

pub fn pll1divdis( self ) -> RegisterField<7, 0x1, 1, 0, Pll1Divdis, Ccucon1_SPEC, RW>

Divider Disable for fPLL1 PLL1DIVDIS. Depending on CCUCON0.CLKSEL this bit selects whether f source1 is half f pll1 .

pub fn i2cdiv(self) -> RegisterField<8, 0xf, 1, 0, I2Cdiv, Ccucon1_SPEC, RW>

I2C Divider Reload Value I2CDIV. The resulting I2C frequency is configured to f I2C xa0 xa0 f SOURCE2 xa0 xa0 I2CDIV for the allowed configurations. For I2CDIV xa0 xa0 0000 B the clock is shut off.

pub fn mscdiv(self) -> RegisterField<16, 0xf, 1, 0, Mscdiv, Ccucon1_SPEC, RW>

MSC Divider Reload Value MSCDIV. The resulting MSC frequency is configured to f MSC 160 160 f SOURCEMSC 160 160 MSCDIV for the allowed configurations. For MSCDIV 160 160 0000 B the clock is shut off.

pub fn clkselmsc( self ) -> RegisterField<20, 0x3, 1, 0, Clkselmsc, Ccucon1_SPEC, RW>

Clock Selection for MSC CLKSELMSC. This bit field defines the clock source that is used for the clock generation of f SOURCEMSC . For switching between two non zero configurations the following sequence has to be applied First step is to switch to 00 B . Second step is to switch to the new target configuration.

pub fn qspidiv(self) -> RegisterField<24, 0xf, 1, 0, Qspidiv, Ccucon1_SPEC, RW>

QSPI Divider Reload Value QSPIDIV. The resulting QSPI frequency is configured to f QSPI 160 160 f SOURCEQSPI 160 160 QSPIDIV for the allowed configurations. For QSPIDIV 160 160 0000 B the clock is shut off.

pub fn clkselqspi( self ) -> RegisterField<28, 0x3, 1, 0, Clkselqspi, Ccucon1_SPEC, RW>

Clock Selection for QSPI CLKSELQSPI. This bit field defines the clock source that is used for the clock generation of f SOURCEQSPI . For switching between two non zero configurations the following sequence has to be applied First step is to switch to 00 B . Second step is to switch to the new target configuration.

pub fn lck(self) -> RegisterField<31, 0x1, 1, 0, Lck, Ccucon1_SPEC, R>

Lock Status LCK. This bit indicates if the register can be updated with a new value or if the register is locked and a write action from the bus side has no effect. The lock bit is set when at least one bit field is changed and released when this change is executed.