Type Definition tc37xpd::scu::Ccucon3

pub type Ccucon3 = RegValueT<Ccucon3_SPEC>;

CCU Clock Control Register 3 resetvalue={System Reset:0x0}

Implementations

impl Ccucon3

pub fn pll0monen( self ) -> RegisterField<0, 0x1, 1, 0, Pll0Monen, Ccucon3_SPEC, RW>

PLL0 Clock Monitor Enable PLL0MONEN

pub fn pll1monen( self ) -> RegisterField<1, 0x1, 1, 0, Pll1Monen, Ccucon3_SPEC, RW>

PLL1 Clock Monitor Enable PLL1MONEN

pub fn pll2monen( self ) -> RegisterField<2, 0x1, 1, 0, Pll2Monen, Ccucon3_SPEC, RW>

PLL2 Clock Monitor Enable PLL2MONEN

pub fn spbmonen(self) -> RegisterField<3, 0x1, 1, 0, Spbmonen, Ccucon3_SPEC, RW>

SPB Clock Monitor Enable SPBMONEN

pub fn backmonen( self ) -> RegisterField<4, 0x1, 1, 0, Backmonen, Ccucon3_SPEC, RW>

Backup Clock Monitor Enable BACKMONEN

pub fn pll0montst( self ) -> RegisterField<8, 0x1, 1, 0, Pll0Montst, Ccucon3_SPEC, RW>

PLL0 Clock Monitor Test PLL0MONTST. The test enable bit is not a direct trigger for the alarm but an inhibit for the clock to be monitored. This is to test the monitoring logic itself as well.

pub fn pll1montst( self ) -> RegisterField<9, 0x1, 1, 0, Pll1Montst, Ccucon3_SPEC, RW>

PLL1 Clock Monitor Test PLL1MONTST. The test enable bit is not a direct trigger for the alarm but an inhibit for the clock to be monitored. This is to test the monitoring logic itself as well.

pub fn pll2montst( self ) -> RegisterField<10, 0x1, 1, 0, Pll2Montst, Ccucon3_SPEC, RW>

PLL2 Clock Monitor Test PLL2MONTST. The test enable bit is not a direct trigger for the alarm but an inhibit for the clock to be monitored. This is to test the monitoring logic itself as well.

pub fn spbmontst( self ) -> RegisterField<11, 0x1, 1, 0, Spbmontst, Ccucon3_SPEC, RW>

SPB Clock Monitor Test SPBMONTST. The test enable bit is not a direct trigger for the alarm but an inhibit for the clock to be monitored. This is to test the monitoring logic itself as well.

pub fn backmontst( self ) -> RegisterField<12, 0x1, 1, 0, Backmontst, Ccucon3_SPEC, RW>

Backup Clock Monitor Test BACKMONTST. The test enable bit is not a direct trigger for the alarm but an inhibit for the clock to be monitored. This is to test the monitoring logic itself as well.

pub fn up(self) -> RegisterField<30, 0x1, 1, 0, Up, Ccucon3_SPEC, W>

Update Request UP. Setting this bit will request an update for CCUCON3 and CCUCON4. Only one UP bit must be set for either CCUCON3 or CCUCON4. This bit always reads as zero.

pub fn lck(self) -> RegisterField<31, 0x1, 1, 0, Lck, Ccucon3_SPEC, R>

Lock Status LCK. This bit indicates if the register can be updated with a new value or if the register is locked and a write action from the bus side has no effect. The lock bit is set when an update of CCUCON3 4 has been requested and released when the update is complete.