Type Definition tc37xpd::scu::Ccucon5

pub type Ccucon5 = RegValueT<Ccucon5_SPEC>;

CCU Clock Control Register 5 resetvalue={System Reset:0x30,System Reset:0x30}

Implementations

impl Ccucon5

pub fn gethdiv(self) -> RegisterField<0, 0xf, 1, 0, Gethdiv, Ccucon5_SPEC, RW>

GETH Divider Reload Value GETHDIV. The resulting GETH frequency is configured to f GETH xa0 xa0 f source0 xa0 xa0 GETHDIV for the allowed configurations. For GETHDIV xa0 xa0 0000 B the clock is shut off. f source0 could be configured either to f PLL0 CLKSEL xa0 xa0 01 B or f BACK CLKSEL xa0 xa0 00 B GETHDIV must be enabled 0 during an application reset to allow firmware related installation tasks.

pub fn mcanhdiv(self) -> RegisterField<4, 0xf, 1, 0, Mcanhdiv, Ccucon5_SPEC, RW>

MCANH Divider Reload Value MCANHDIV. The resulting MCANH frequency is configured to f MCANH f SOURCE0 MCANHDIV for the allowed configurations. For MCANHDIV 0000 B the clock is shut off.

pub fn up(self) -> RegisterField<30, 0x1, 1, 0, Up, Ccucon5_SPEC, W>

Update Request UP. Setting this bit will request an update for CCUCON0 and CCUCON5. Only one UP bit must be set either CCUCON0 or CCUCON5. This bit always reads as zero.

pub fn lck(self) -> RegisterField<31, 0x1, 1, 0, Lck, Ccucon5_SPEC, R>

Lock Status LCK. This bit indicates if the register can be updated with a new value or if the register is locked and a write action from the bus side has no effect. The lock bit is set when an update of CCUCON0 5 has been requested and released when the update is complete.