Type Definition tc37xpd::scu::Chipid

pub type Chipid = RegValueT<Chipid_SPEC>;

Chip Identification Register resetvalue={System Reset:0x0,CFS Value:0x0}

Implementations

impl Chipid

pub fn chrev(self) -> RegisterField<0, 0x3f, 1, 0, u8, Chipid_SPEC, R>

Chip Revision Number CHREV. This bit field indicates the revision number of the AurixPlus Platform device. The value of this bit field is defined in the product Data Sheet. Bits 3 0 are connected to the layer prog inputs 3 0 . Bits 3 0 are used to indicate the steps. These updates can be done with any metal fix or FW ROM change. Bits 5 4 are hard wired on top level. Bits 5 4 define the major silicon design steps A B C D … . These bits can be changed only with a major redesign.

pub fn chtec(self) -> RegisterField<6, 0x3, 1, 0, Chtec, Chipid_SPEC, R>

Chip Family CHTEC. These bits indicate the product family and are changed only with a redesign. Hard wired on top level.

pub fn chpk(self) -> RegisterField<8, 0xf, 1, 0, Chpk, Chipid_SPEC, RW>

Chip Package CHPK. These bits indicate the package. For further details refer to the Product Datasheet Updated by SSW from config sector. Use future variant codes for downconfigured silicon

pub fn chid(self) -> RegisterField<12, 0xf, 1, 0, Chid, Chipid_SPEC, RW>

Chip Product CHID. These bits indicate the base product. For further details refer to the Product Datasheet Updated by SSW from config sector.

pub fn eea(self) -> RegisterField<16, 0x1, 1, 0, Eea, Chipid_SPEC, R>

Emulation or ADAS Extension Available EEA. Indicates if the emulation or ADAS extension hardware is available or not.

pub fn ucode(self) -> RegisterField<17, 0x7f, 1, 0, u8, Chipid_SPEC, RW>

µCode Version UCODE. This bit field displays the Version X.Y of the flash µCode.

pub fn fsize(self) -> RegisterField<24, 0xf, 1, 0, Fsize, Chipid_SPEC, RW>

Program Flash Size FSIZE. This bit field indicates available program flash size for this device. For more details see Product Datasheet.

pub fn vart(self) -> RegisterField<28, 0x7, 1, 0, u8, Chipid_SPEC, RW>

Variant VART. This bit field is used for variant identification. It is used to identify product variants with non standard temperature profile max frequency package pitch or customer feature sets. Note that variants do not include HSM availability Flash Size or Emulation availability options which are handled by the SEC FSIZE and EEA fields respectively. For coding details see Product Datasheet

pub fn sec(self) -> RegisterField<31, 0x1, 1, 0, Sec, Chipid_SPEC, RW>

Security Device Available SEC. This bit field is updated by SSW from config sector. This bit field indicates whether the product has a Hardware Security Module