Type Definition tc37xpd::scu::Clkdivctrl0

pub type Clkdivctrl0 = RegValueT<Clkdivctrl0_SPEC>;

Clock Division Control Register 0 resetvalue={System Reset:0x0}

Implementations

impl Clkdivctrl0

pub fn cnten(self) -> RegisterField<0, 0x1, 1, 0, Cnten, Clkdivctrl0_SPEC, RW>

Counter Enable CNTEN. This field specifies if the CCU clock counter logic for DfT Test is enabled for internal frequency measurement.

pub fn trigsel( self ) -> RegisterField<2, 0x1f, 1, 0, Trigsel, Clkdivctrl0_SPEC, RW>

Trigger Selection TRIGSEL. This field allows to select between various trigger signals defining start and stop time for the next counter measurement period.

pub fn smondiv( self ) -> RegisterField<7, 0x1, 1, 0, Smondiv, Clkdivctrl0_SPEC, RW>

SMON3 Pre Division Factor Selection SMONDIV. This bit directly selects the pre division factor for the ring oscillation output frequency of all SMON3 macro cells Reg VT and High VT . If TEST TRC4.SMONSEL lt 1 or 2 gt is set to a different value than 0000 this bit has no meaning because in this case TCU directly selects the SMON3 output division factor via TEST TRC4.SMONDIV

pub fn outenp(self) -> RegisterField<8, 0x1, 1, 0, Outenp, Clkdivctrl0_SPEC, RW>

Output Enabling Protection OUTENP. Protection bit for automatic output enabling in case of TRIGSEL 11111

pub fn clkdivsel( self ) -> RegisterField<9, 0x3, 1, 0, Clkdivsel, Clkdivctrl0_SPEC, RW>

Output Select for PLL Clock Divider CLKDIVSEL. This bit allows to select which of PLL divider counter inside CCU will be forwarded to vtp24 TDO GPIO in case of continuous counting mode. TRIGSEL 11111 .

pub fn cntsel( self ) -> RegisterField<12, 0x3, 1, 0, Cntsel, Clkdivctrl0_SPEC, RW>

Counter Output Selection CNTSEL. This register selects which counter value is readable through the CLKDIVCTRL1.COUNTVAL bitfield Note Please be aware that after changing this bit field it will take some clock cycles before CLKDIVCTRL1.COUNTVAL is updated.

pub fn start(self) -> RegisterFieldBool<14, 1, 0, Clkdivctrl0_SPEC, RW>

Start CCU Clock Counting START. If written high this bit triggers the automatic accumulation of PLL clock pulses during a specific period defined through TRIGSEL . After each period counter will restart until the STOP bit has been set. If START and STOP bits are set to high concurrently counter will run for exactly one TRIGSEL period. When written with a high value this bit will reset automatically as soon counter active signal from CCU has been set to high. As long as CLKDIVCTRL1.BUSY is high any write access to this bit is ignored. In case of continuous counting TRIGSEL 11111 this bit will have no influence. To avoid transitory configuration values in CCU while counter is active TRIGSEL should not be changed concurrently while setting this bit to 1 .

pub fn stop(self) -> RegisterFieldBool<15, 1, 0, Clkdivctrl0_SPEC, RW>

Stop CCU Clock Counting STOP. If written high this bit stops the automatic accumulation of PLL clock pulses after occurrence of the next trigger signal defined through TRIGSEL. After written with a high value this bit will be synchronously reset in case CLKDIVCTRL1.BUSY and START bits are both low i.e. If START and STOP bits are set to 1 concurrently while counter is idle the automatic reset of STOP must not happen before counter has been active for one TRIGSEL period . In case of continuous counting TRIGSEL 1111 this bit will have no influence.

pub fn smonctr( self ) -> RegisterField<16, 0x7, 1, 0, Smonctr, Clkdivctrl0_SPEC, RW>

Speedmon SMONCTR. These registers enable and control the SMON3 Ring Oscillator macro cells for all reg VT and high VT types. If TEST TRC4.SMONSEL lt 1 or 2 gt is set to a different value than 0000 these bits have no meaning because TCU directly controls the SMON cells via TEST TRC4.SMONCTR in that case.

pub fn prescale( self ) -> RegisterField<19, 0x1fff, 1, 0, Prescale, Clkdivctrl0_SPEC, RW>

Pre scaling factor PRESCALE. This register allows to determine the pre scaling factor for the selected trigger source. Please regard that the pre scaler control vector consists of 14 bits from which only the 13 MSBs are controllable through this register. Bit 0 is not controllable and always tied to 0 .The following settings are supported