Type Definition tc37xpd::scu::Dtscbgoctrl

pub type Dtscbgoctrl = RegValueT<Dtscbgoctrl_SPEC>;

Core Die Temperature Sensor Bandgap Control Register resetvalue={Cold PowerOn Reset:0x40,CFS Value:0x40}

Implementations

impl Dtscbgoctrl

pub fn vbgtrim(self) -> RegisterField<0, 0x7f, 1, 0, u8, Dtscbgoctrl_SPEC, RW>

DTSC Bandgap Trim Value VBGTRIM. This bit field contains information about trimming of the DTSC bandgap.The Bandgap voltage is measured via Analog Test Bus at 125 176 C and the value is trimmed so that the measurement corresponds to 1.13V. A settling time of 30 160 us need to be waited after a new trim value is programmed. trimvpbg trimbg i VBGTRIM VBGPTRIM signed value

pub fn vbgptrim(self) -> RegisterField<21, 0x1f, 1, 0, u8, Dtscbgoctrl_SPEC, RW>

DTSC Bandgap Signed Trim Value VBGPTRIM. This bit field allows device individual trimming of the bandgap trim value in production. The first bit is sign information. The range is 16 to 15.