Type Definition tc37xpd::scu::Dtsclim

pub type Dtsclim = RegValueT<Dtsclim_SPEC>;

Core Die Temperature Sensor Limit Register resetvalue={Application Reset:0x0CD806D6}

Implementations

impl Dtsclim

pub fn lower(self) -> RegisterField<0, 0xfff, 1, 0, u16, Dtsclim_SPEC, RW>

DTSC Lower Limit LOWER. This bit field defines the lower limit of the DTSC temperature check. The DTSC measurement result is compared against this value and if the measurement result is less than or equal to the configured LOWER bitfield value flag LLU is set.

pub fn bgpok(self) -> RegisterField<13, 0x1, 1, 0, Bgpok, Dtsclim_SPEC, R>

DTSC Bandgap OK. This bitfield indicates that the bandgap reference for the Core Die Temperature Sensor DTSC is available and ok.

pub fn en(self) -> RegisterField<14, 0x1, 1, 0, En, Dtsclim_SPEC, RW>

DTSC Enable. This bitfield enables the Core Die Temperature Sensor DTSC . The bitfield is reset on an application reset.

pub fn llu(self) -> RegisterField<15, 0x1, 1, 0, Llu, Dtsclim_SPEC, RW>

DTSC Lower Limit Underflow LLU. When this bit is set the related SMU DTSC alarm trigger is generated. This bit has to be written with zero in order to clear it. Writing a one has no effect. This bit is set when a DTSC measurement is finished and the result is below the lower limit i.e. DTSCLIM.LOWER .

pub fn upper(self) -> RegisterField<16, 0xfff, 1, 0, u16, Dtsclim_SPEC, RW>

DTSC Upper Limit UPPER. This bit field defines the upper limit of the DTSC temperature check. The DTSC measurement result is compared against this value and if the measurement result is greater than or equal to the configured UPPER bitfield value flag UOF is set.

pub fn inten(self) -> RegisterField<28, 0x1, 1, 0, Inten, Dtsclim_SPEC, RW>

DTSC Interrupt Enable. This bitfield enables the Core Die Temperature Sensor DTSC interrupt. The bitfield is reset on an application reset.

pub fn int(self) -> RegisterField<30, 0x1, 1, 0, Int, Dtsclim_SPEC, RW>

DTSC Interrupt status flag. This bit is set when SMU DTSC interrupt is generated when a DTSC measurement is finished. This bit is cleared by writing a zero. Writing a one has no effect.

pub fn uof(self) -> RegisterField<31, 0x1, 1, 0, Uof, Dtsclim_SPEC, RW>

DTSC Upper Limit Overflow UOF. When this bit is set the related SMU DTSC alarm trigger is generated. This bit has to be written with zero in order to clear it. Writing a one has no effect. This bit is set when a DTSC measurement is finished and the result is exceeding the upper limit i.e. DTSCLIM.UPPER .