Type Definition tc37xpd::scu::EicRi

pub type EicRi = RegValueT<EicRi_SPEC>;

External Input Channel Register 0 resetvalue={Application Reset:0x0}

Implementations

impl EicRi

pub fn exis0(self) -> RegisterField<4, 0x7, 1, 0, Exis0, EicRi_SPEC, RW>

External Input Selection 0 EXIS0. This bit field determines which input line is selected for Input Channel 2i .

pub fn fen0(self) -> RegisterField<8, 0x1, 1, 0, Fen0, EicRi_SPEC, RW>

Falling Edge Enable 0 FEN0. This bit determines if the falling edge of Input Channel 2i is used to set bit INTF 2i .

pub fn ren0(self) -> RegisterField<9, 0x1, 1, 0, Ren0, EicRi_SPEC, RW>

Rising Edge Enable 0 REN0. This bit determines if the rising edge of Input Channel 2 i is used to set bit INTF 2i .

pub fn lden0(self) -> RegisterField<10, 0x1, 1, 0, Lden0, EicRi_SPEC, RW>

Level Detection Enable 0 LDEN0. This bit determines if bit INTF 2i is cleared automatically if an edge of the input Input Channel 2i is detected which has not been selected rising edge with REN0 0 or falling edge with FEN0 0 .

pub fn eien0(self) -> RegisterField<11, 0x1, 1, 0, Eien0, EicRi_SPEC, RW>

External Input Enable 0 EIEN0. This bit enables the generation of a trigger event for request channel 2i e.g. for interrupt generation when a selected edge is detected.

pub fn inp0(self) -> RegisterField<12, 0x7, 1, 0, Inp0, EicRi_SPEC, RW>

Input Node Pointer INP0. This bit field determines the destination output channel for trigger event 2i if enabled by EIEN 2i .

pub fn exis1(self) -> RegisterField<20, 0x7, 1, 0, Exis1, EicRi_SPEC, RW>

External Input Selection 1 EXIS1. This bit field determines which input line is selected for Input Channel 2i 1 .

pub fn fen1(self) -> RegisterField<24, 0x1, 1, 0, Fen1, EicRi_SPEC, RW>

Falling Edge Enable 1 FEN1. This bit determines if the falling edge of Input Channel 2i 1 is used to set bit INTF 2i 1 .

pub fn ren1(self) -> RegisterField<25, 0x1, 1, 0, Ren1, EicRi_SPEC, RW>

Rising Edge Enable 1 REN1. This bit determines if the rising edge of Input Channel 2i 1 is used to set bit INTF 2i 1 .

pub fn lden1(self) -> RegisterField<26, 0x1, 1, 0, Lden1, EicRi_SPEC, RW>

Level Detection Enable 1 LDEN1. This bit determines if bit INTF 2i 1 is cleared automatically if an edge of the input Input Channel 2i 1 is detected which has not been selected rising edge with REN1 0 or falling edge with FEN1 0 .

pub fn eien1(self) -> RegisterField<27, 0x1, 1, 0, Eien1, EicRi_SPEC, RW>

External Input Enable 1 EIEN1. This bit enables the generation of a trigger event for request channel 2i 1 e.g. for interrupt generation when a selected edge is detected.

pub fn inp1(self) -> RegisterField<28, 0x7, 1, 0, Inp1, EicRi_SPEC, RW>

Input Node Pointer INP1. This bit field determines the destination output channel for trigger event 2i 1 if enabled by EIEN 2i 1 .