Type Definition tc37xpd::scu::Eifilt

pub type Eifilt = RegValueT<Eifilt_SPEC>;

External Input Filter Register resetvalue={Application Reset:0x0}

Implementations

impl Eifilt

pub fn filrq0a(self) -> RegisterField<0, 0x1, 1, 0, Filrq0A, Eifilt_SPEC, RW>

Filter Enable for REQ0A FILRQ0A

pub fn filrq5a(self) -> RegisterField<1, 0x1, 1, 0, Filrq5A, Eifilt_SPEC, RW>

Filter Enable for REQ5A FILRQ5A

pub fn filrq2a(self) -> RegisterField<2, 0x1, 1, 0, Filrq2A, Eifilt_SPEC, RW>

Filter Enable for REQ2A FILRQ2A

pub fn filrq3a(self) -> RegisterField<3, 0x1, 1, 0, Filrq3A, Eifilt_SPEC, RW>

Filter Enable for REQ3A FILRQ3A

pub fn filrq0c(self) -> RegisterField<4, 0x1, 1, 0, Filrq0C, Eifilt_SPEC, RW>

Filter Enable for REQ0C FILRQ0C

pub fn filrq1c(self) -> RegisterField<5, 0x1, 1, 0, Filrq1C, Eifilt_SPEC, RW>

Filter Enable for REQ1C FILRQ1C

pub fn filrq3c(self) -> RegisterField<6, 0x1, 1, 0, Filrq3C, Eifilt_SPEC, RW>

Filter Enable for REQ3C FILRQ3C

pub fn filrq2c(self) -> RegisterField<7, 0x1, 1, 0, Filrq2C, Eifilt_SPEC, RW>

Filter Enable for REQ2C FILRQ2C

pub fn filrq4a(self) -> RegisterField<8, 0x1, 1, 0, Filrq4A, Eifilt_SPEC, RW>

Filter Enable for REQ4A FILRQ4A

pub fn filrq6a(self) -> RegisterField<9, 0x1, 1, 0, Filrq6A, Eifilt_SPEC, RW>

Filter Enable for REQ6A FILRQ6A

pub fn filrq1a(self) -> RegisterField<10, 0x1, 1, 0, Filrq1A, Eifilt_SPEC, RW>

Filter Enable for REQ1A FILRQ1A

pub fn filrq7a(self) -> RegisterField<11, 0x1, 1, 0, Filrq7A, Eifilt_SPEC, RW>

Filter Enable for REQ7A FILRQ7A

pub fn filrq6d(self) -> RegisterField<12, 0x1, 1, 0, Filrq6D, Eifilt_SPEC, RW>

Filter Enable for REQ6D FILRQ6D

pub fn filrq4d(self) -> RegisterField<13, 0x1, 1, 0, Filrq4D, Eifilt_SPEC, RW>

Filter Enable for REQ4D FILRQ4D

pub fn filrq2b(self) -> RegisterField<14, 0x1, 1, 0, Filrq2B, Eifilt_SPEC, RW>

Filter Enable for REQ2B FILRQ2B

pub fn filrq3b(self) -> RegisterField<15, 0x1, 1, 0, Filrq3B, Eifilt_SPEC, RW>

Filter Enable for REQ3B FILRQ3B

pub fn filrq7c(self) -> RegisterField<16, 0x1, 1, 0, Filrq7C, Eifilt_SPEC, RW>

Filter Enable for REQ7C FILRQ7C

pub fn filtdiv(self) -> RegisterField<24, 0xf, 1, 0, u8, Eifilt_SPEC, RW>

Digital Glitch Filter Clock Predivider FILTDIV. This field controls a predivider to generate the digital filter sample clock T filt T spb FILTDIV A value of zero in this register disables all glitch filtering.

pub fn depth(self) -> RegisterField<28, 0xf, 1, 0, u8, Eifilt_SPEC, RW>

Digital Glitch Filter Depth DEPTH. DEPTH determines the number of port input samples considered in the calculation of the floating average digital filter output for all enabled FLRQ filters. A value of zero in this register disables all glitch filtering.