Type Definition tc37xpd::scu::Eisr

pub type Eisr = RegValueT<Eisr_SPEC>;

ENDINIT Timeout Counter Status Register resetvalue={Application Reset:0x0FFFC0000}

Implementations

impl Eisr

pub fn ae(self) -> RegisterField<0, 0x1, 1, 0, Ae, Eisr_SPEC, R>

EICON0 Access Error Status Flag AE. This bit is set when an illegal Password Access or Modify Access to register EICON0 was attempted. This bit is only cleared on a valid EICON0.ENDINIT Modify Access

pub fn oe(self) -> RegisterField<1, 0x1, 1, 0, Oe, Eisr_SPEC, R>

EI Timeout Overflow Error Status Flag OE. This bit is set when EISR.TIM overflows from FFFF H to FFFC H . This bit is only cleared on a valid EICON0 Modify Access.

pub fn is0(self) -> RegisterField<2, 0x1, 1, 0, Is0, Eisr_SPEC, R>

EI Timeout Input Clock Status IS1 IS0. Bit IS0 and IS1 should be programmed together. These bits indicate the current ENDINIT Timeout Counter frequency.

pub fn ds(self) -> RegisterField<3, 0x1, 1, 0, Ds, Eisr_SPEC, R>

EI Timeout Enable Disable Status Flag DS

pub fn to(self) -> RegisterField<4, 0x1, 1, 0, To, Eisr_SPEC, R>

EI Time Out Mode Flag TO

pub fn is1(self) -> RegisterField<5, 0x1, 1, 0, Is1, Eisr_SPEC, R>

EI Timeout Input Clock Status IS1 IS0. Bit IS0 and IS1 should be programmed together. These bits indicate the current ENDINIT Timeout Counter frequency.

pub fn tim(self) -> RegisterField<16, 0xffff, 1, 0, u16, Eisr_SPEC, R>

Timer Value TIM. Reflects the current content of the ENDINIT Timeout Counter. Only bits 17 and 16 are implemented in EISR. Others return 8216 1 8217