Type Definition tc37xpd::scu::Extcon

pub type Extcon = RegValueT<Extcon_SPEC>;

External Clock Control Register resetvalue={System Reset:0x0}

Implementations

impl Extcon

pub fn en0(self) -> RegisterField<0, 0x1, 1, 0, En0, Extcon_SPEC, RW>

External Clock Enable for EXTCLK0 EN0. If the generation of the external clock signal is disabled the signal is tied to zero. This bit field can be overruled in Test Mode via the TCU.

pub fn sel0(self) -> RegisterField<2, 0xf, 1, 0, Sel0, Extcon_SPEC, RW>

External Clock Select for EXTCLK0 SEL0. This bit field defines the clock source that is selected as output for pin EXTCLK0. This bit field can be overruled in Test Mode via the TCU.

pub fn en1(self) -> RegisterField<16, 0x1, 1, 0, En1, Extcon_SPEC, RW>

External Clock Enable for EXTCLK1 EN1. If the generation of the external clock signal is disabled the signal is tied to zero. This bit can be overruled in Test Mode via the TCU.

pub fn nsel(self) -> RegisterField<17, 0x1, 1, 0, Nsel, Extcon_SPEC, RW>

Negation Selection NSEL. This bit can be overruled in Test Mode via the TCU.

pub fn sel1(self) -> RegisterField<18, 0xf, 1, 0, Sel1, Extcon_SPEC, RW>

External Clock Select for EXTCLK1 SEL1. This bit field defines the clock source that is selected as the output for pin EXTCLK1. This bit field can be overruled in Test Mode via the TCU.

pub fn div1(self) -> RegisterField<24, 0xff, 1, 0, u8, Extcon_SPEC, RW>

External Clock Divider for EXTCLK1 DIV1. This value defines the reload value of the divider that generates f OUT out of f SPB f OUT f SPB DIV1 1 . The divider itself is cleared each time bit EN1 is cleared.