Type Definition tc37xpd::scu::Fmr

pub type Fmr = RegValueT<Fmr_SPEC>;

Flag Modification Register resetvalue={Application Reset:0x0}

Implementations

impl Fmr

pub fn fs0(self) -> RegisterField<0, 0x1, 1, 0, Fs0, Fmr_SPEC, W>

Set Flag INTFx for Channel 7 FS7. Setting this bit will set the corresponding bit INTFx in register EIFR. Reading this bit always delivers a 0. If both FSx and FCx are set in the same access then the bit x in register EIFR is not modified.

pub fn fs1(self) -> RegisterField<1, 0x1, 1, 0, Fs1, Fmr_SPEC, W>

Set Flag INTFx for Channel 7 FS7. Setting this bit will set the corresponding bit INTFx in register EIFR. Reading this bit always delivers a 0. If both FSx and FCx are set in the same access then the bit x in register EIFR is not modified.

pub fn fs2(self) -> RegisterField<2, 0x1, 1, 0, Fs2, Fmr_SPEC, W>

Set Flag INTFx for Channel 7 FS7. Setting this bit will set the corresponding bit INTFx in register EIFR. Reading this bit always delivers a 0. If both FSx and FCx are set in the same access then the bit x in register EIFR is not modified.

pub fn fs3(self) -> RegisterField<3, 0x1, 1, 0, Fs3, Fmr_SPEC, W>

Set Flag INTFx for Channel 7 FS7. Setting this bit will set the corresponding bit INTFx in register EIFR. Reading this bit always delivers a 0. If both FSx and FCx are set in the same access then the bit x in register EIFR is not modified.

pub fn fs4(self) -> RegisterField<4, 0x1, 1, 0, Fs4, Fmr_SPEC, W>

Set Flag INTFx for Channel 7 FS7. Setting this bit will set the corresponding bit INTFx in register EIFR. Reading this bit always delivers a 0. If both FSx and FCx are set in the same access then the bit x in register EIFR is not modified.

pub fn fs5(self) -> RegisterField<5, 0x1, 1, 0, Fs5, Fmr_SPEC, W>

Set Flag INTFx for Channel 7 FS7. Setting this bit will set the corresponding bit INTFx in register EIFR. Reading this bit always delivers a 0. If both FSx and FCx are set in the same access then the bit x in register EIFR is not modified.

pub fn fs6(self) -> RegisterField<6, 0x1, 1, 0, Fs6, Fmr_SPEC, W>

Set Flag INTFx for Channel 7 FS7. Setting this bit will set the corresponding bit INTFx in register EIFR. Reading this bit always delivers a 0. If both FSx and FCx are set in the same access then the bit x in register EIFR is not modified.

pub fn fs7(self) -> RegisterField<7, 0x1, 1, 0, Fs7, Fmr_SPEC, W>

Set Flag INTFx for Channel 7 FS7. Setting this bit will set the corresponding bit INTFx in register EIFR. Reading this bit always delivers a 0. If both FSx and FCx are set in the same access then the bit x in register EIFR is not modified.

pub fn fc0(self) -> RegisterField<16, 0x1, 1, 0, Fc0, Fmr_SPEC, W>

Clear Flag INTFx for Channel 7 FC7. Setting this bit will clear the corresponding bit INTFx in register EIFR. Reading this bit always delivers a 0. If both FSx and FCx are set in the same access then the bit x in register EIFR is not modified.

pub fn fc1(self) -> RegisterField<17, 0x1, 1, 0, Fc1, Fmr_SPEC, W>

Clear Flag INTFx for Channel 7 FC7. Setting this bit will clear the corresponding bit INTFx in register EIFR. Reading this bit always delivers a 0. If both FSx and FCx are set in the same access then the bit x in register EIFR is not modified.

pub fn fc2(self) -> RegisterField<18, 0x1, 1, 0, Fc2, Fmr_SPEC, W>

Clear Flag INTFx for Channel 7 FC7. Setting this bit will clear the corresponding bit INTFx in register EIFR. Reading this bit always delivers a 0. If both FSx and FCx are set in the same access then the bit x in register EIFR is not modified.

pub fn fc3(self) -> RegisterField<19, 0x1, 1, 0, Fc3, Fmr_SPEC, W>

Clear Flag INTFx for Channel 7 FC7. Setting this bit will clear the corresponding bit INTFx in register EIFR. Reading this bit always delivers a 0. If both FSx and FCx are set in the same access then the bit x in register EIFR is not modified.

pub fn fc4(self) -> RegisterField<20, 0x1, 1, 0, Fc4, Fmr_SPEC, W>

Clear Flag INTFx for Channel 7 FC7. Setting this bit will clear the corresponding bit INTFx in register EIFR. Reading this bit always delivers a 0. If both FSx and FCx are set in the same access then the bit x in register EIFR is not modified.

pub fn fc5(self) -> RegisterField<21, 0x1, 1, 0, Fc5, Fmr_SPEC, W>

Clear Flag INTFx for Channel 7 FC7. Setting this bit will clear the corresponding bit INTFx in register EIFR. Reading this bit always delivers a 0. If both FSx and FCx are set in the same access then the bit x in register EIFR is not modified.

pub fn fc6(self) -> RegisterField<22, 0x1, 1, 0, Fc6, Fmr_SPEC, W>

Clear Flag INTFx for Channel 7 FC7. Setting this bit will clear the corresponding bit INTFx in register EIFR. Reading this bit always delivers a 0. If both FSx and FCx are set in the same access then the bit x in register EIFR is not modified.

pub fn fc7(self) -> RegisterField<23, 0x1, 1, 0, Fc7, Fmr_SPEC, W>

Clear Flag INTFx for Channel 7 FC7. Setting this bit will clear the corresponding bit INTFx in register EIFR. Reading this bit always delivers a 0. If both FSx and FCx are set in the same access then the bit x in register EIFR is not modified.