Type Definition tc37xpd::scu::IgcRj

pub type IgcRj = RegValueT<IgcRj_SPEC>;

Flag Gating Register 0 resetvalue={Application Reset:0x0}

Implementations

impl IgcRj

pub fn ipen00(self) -> RegisterField<0, 0x1, 1, 0, Ipen00, IgcRj_SPEC, RW>

Flag Pattern Enable for Channel 0 IPEN07. Bit IPEN0x determines if the flag INTFx of channel x takes part in the pattern detection for the gating of the requests for the output signals GOUT 2 j and IOUT 2 j .

pub fn ipen01(self) -> RegisterField<1, 0x1, 1, 0, Ipen01, IgcRj_SPEC, RW>

Flag Pattern Enable for Channel 0 IPEN07. Bit IPEN0x determines if the flag INTFx of channel x takes part in the pattern detection for the gating of the requests for the output signals GOUT 2 j and IOUT 2 j .

pub fn ipen02(self) -> RegisterField<2, 0x1, 1, 0, Ipen02, IgcRj_SPEC, RW>

Flag Pattern Enable for Channel 0 IPEN07. Bit IPEN0x determines if the flag INTFx of channel x takes part in the pattern detection for the gating of the requests for the output signals GOUT 2 j and IOUT 2 j .

pub fn ipen03(self) -> RegisterField<3, 0x1, 1, 0, Ipen03, IgcRj_SPEC, RW>

Flag Pattern Enable for Channel 0 IPEN07. Bit IPEN0x determines if the flag INTFx of channel x takes part in the pattern detection for the gating of the requests for the output signals GOUT 2 j and IOUT 2 j .

pub fn ipen04(self) -> RegisterField<4, 0x1, 1, 0, Ipen04, IgcRj_SPEC, RW>

Flag Pattern Enable for Channel 0 IPEN07. Bit IPEN0x determines if the flag INTFx of channel x takes part in the pattern detection for the gating of the requests for the output signals GOUT 2 j and IOUT 2 j .

pub fn ipen05(self) -> RegisterField<5, 0x1, 1, 0, Ipen05, IgcRj_SPEC, RW>

Flag Pattern Enable for Channel 0 IPEN07. Bit IPEN0x determines if the flag INTFx of channel x takes part in the pattern detection for the gating of the requests for the output signals GOUT 2 j and IOUT 2 j .

pub fn ipen06(self) -> RegisterField<6, 0x1, 1, 0, Ipen06, IgcRj_SPEC, RW>

Flag Pattern Enable for Channel 0 IPEN07. Bit IPEN0x determines if the flag INTFx of channel x takes part in the pattern detection for the gating of the requests for the output signals GOUT 2 j and IOUT 2 j .

pub fn ipen07(self) -> RegisterField<7, 0x1, 1, 0, Ipen07, IgcRj_SPEC, RW>

Flag Pattern Enable for Channel 0 IPEN07. Bit IPEN0x determines if the flag INTFx of channel x takes part in the pattern detection for the gating of the requests for the output signals GOUT 2 j and IOUT 2 j .

pub fn geen0(self) -> RegisterField<13, 0x1, 1, 0, Geen0, IgcRj_SPEC, RW>

Generate Event Enable 0 GEEN0. Bit GEEN0 enables the generation of a trigger event for output channel 2 j when the result of the pattern detection changes. When using this feature a trigger e.g. for an interrupt is generated during the first clock cycle when a pattern is detected or when it is no longer detected.

pub fn igp0(self) -> RegisterField<14, 0x3, 1, 0, Igp0, IgcRj_SPEC, RW>

Interrupt Gating Pattern 0 IGP0. In each register IGCRj bit field IGP0 determines how the pattern detection influences the output lines GOUT 2j and IOUT 2j .

pub fn ipen10(self) -> RegisterField<16, 0x1, 1, 0, Ipen10, IgcRj_SPEC, RW>

Interrupt Pattern Enable for Channel 1 IPEN17. Bit IPEN 2j 1 x determines if the flag INTFx of channel 2j 1 takes part in the pattern detection for the gating of the requests for the output signals GOUT 2j 1 and IOUT 2j 1 .

pub fn ipen11(self) -> RegisterField<17, 0x1, 1, 0, Ipen11, IgcRj_SPEC, RW>

Interrupt Pattern Enable for Channel 1 IPEN17. Bit IPEN 2j 1 x determines if the flag INTFx of channel 2j 1 takes part in the pattern detection for the gating of the requests for the output signals GOUT 2j 1 and IOUT 2j 1 .

pub fn ipen12(self) -> RegisterField<18, 0x1, 1, 0, Ipen12, IgcRj_SPEC, RW>

Interrupt Pattern Enable for Channel 1 IPEN17. Bit IPEN 2j 1 x determines if the flag INTFx of channel 2j 1 takes part in the pattern detection for the gating of the requests for the output signals GOUT 2j 1 and IOUT 2j 1 .

pub fn ipen13(self) -> RegisterField<19, 0x1, 1, 0, Ipen13, IgcRj_SPEC, RW>

Interrupt Pattern Enable for Channel 1 IPEN17. Bit IPEN 2j 1 x determines if the flag INTFx of channel 2j 1 takes part in the pattern detection for the gating of the requests for the output signals GOUT 2j 1 and IOUT 2j 1 .

pub fn ipen14(self) -> RegisterField<20, 0x1, 1, 0, Ipen14, IgcRj_SPEC, RW>

Interrupt Pattern Enable for Channel 1 IPEN17. Bit IPEN 2j 1 x determines if the flag INTFx of channel 2j 1 takes part in the pattern detection for the gating of the requests for the output signals GOUT 2j 1 and IOUT 2j 1 .

pub fn ipen15(self) -> RegisterField<21, 0x1, 1, 0, Ipen15, IgcRj_SPEC, RW>

Interrupt Pattern Enable for Channel 1 IPEN17. Bit IPEN 2j 1 x determines if the flag INTFx of channel 2j 1 takes part in the pattern detection for the gating of the requests for the output signals GOUT 2j 1 and IOUT 2j 1 .

pub fn ipen16(self) -> RegisterField<22, 0x1, 1, 0, Ipen16, IgcRj_SPEC, RW>

Interrupt Pattern Enable for Channel 1 IPEN17. Bit IPEN 2j 1 x determines if the flag INTFx of channel 2j 1 takes part in the pattern detection for the gating of the requests for the output signals GOUT 2j 1 and IOUT 2j 1 .

pub fn ipen17(self) -> RegisterField<23, 0x1, 1, 0, Ipen17, IgcRj_SPEC, RW>

Interrupt Pattern Enable for Channel 1 IPEN17. Bit IPEN 2j 1 x determines if the flag INTFx of channel 2j 1 takes part in the pattern detection for the gating of the requests for the output signals GOUT 2j 1 and IOUT 2j 1 .

pub fn geen1(self) -> RegisterField<29, 0x1, 1, 0, Geen1, IgcRj_SPEC, RW>

Generate Event Enable 1 GEEN1. Bit GEEN1 enables the generation of a trigger event for output channel 2j 1 when the result of the pattern detection changes. When using this feature a trigger e.g. for an interrupt is generated during the first clock cycle when a pattern is detected or when it is no longer detected.

pub fn igp1(self) -> RegisterField<30, 0x3, 1, 0, Igp1, IgcRj_SPEC, RW>

Interrupt Gating Pattern 1 IGP1. In each register IGCRj bit field IGP1 determines how the pattern detection influences the output lines GOUT 2j 1 and IOUT 2j 1 .