Type Definition tc37xpd::scu::Iocr

pub type Iocr = RegValueT<Iocr_SPEC>;

Input Output Control Register resetvalue={System Reset:0x0E0}

Implementations

impl Iocr

pub fn pc0(self) -> RegisterField<4, 0xf, 1, 0, u8, Iocr_SPEC, RW>

Control for ESR0 Pin PC0. This bit field defines the ESR0 functionality according to the coding tables.

pub fn pc1(self) -> RegisterField<12, 0xf, 1, 0, u8, Iocr_SPEC, RW>

Control for ESR1 Pin PC1. This bit field defines the ESR1 functionality according to the coding tables. The reset value of SCU IOCR.PC1 is influenced by HWCFG6 and PMSWCR5.TRISTREQ. When a cold reset is activated and HWCFG6 1 then PC1 is reset to 20H and ESR1 will have input pull up mode. If HWCFG6 0 then PC1 is reset to 00H and ESR1 will have tri state mode. PC1 and the ESR1 reset state can also be configured by software with the PMSWCR5.TRISTREQ bit. PMSWCR5.TRISTREQ is not affected by warm reset or wake up from standby so the IOCR.PC1 reset value is configured as per the state of the TRISTREQ bit prior to the warm reset