Type Definition tc37xpd::scu::Lbistctrl0

pub type Lbistctrl0 = RegValueT<Lbistctrl0_SPEC>;

Logic BIST Control 0 Register resetvalue={System Reset:0x0,CFS Value:0x400}

Implementations

impl Lbistctrl0

pub fn lbistreq(self) -> RegisterFieldBool<0, 1, 0, Lbistctrl0_SPEC, W>

LBIST Request LBISTREQ. If written high this bit requests the execution of an automatic scan test procedure. The request will only be approved if LBISTCTRL0 .LBISTDONE bit reflects a 8217 0 8217 value i.e. no LBIST procedure was triggered since the last power on reset or LBIST controller has been restarted through the LBISTCTRL0 .LBISTRES bit . If read this bit always returns a 8216 0 8217 . This bit shall be implemented in a safety relevant way to avoid unintended activation of LBIST during application. LBIST execution time depends on the number of scan loads as defined in the PATTERNS field.

pub fn lbistres(self) -> RegisterFieldBool<1, 1, 0, Lbistctrl0_SPEC, W>

LBIST Reset LBISTRES. If written high this bit synchronously brings back the LBIST controller to its initial Reset Idle state and also clears the stored MISR signature to allow another execution from CPU side. As a consequence the LBISTCTRL0 .LBISTDONE and LBISTCTRL3 .SIGNATURE bits will be set to 8217 0 8217 . If read this bit always returns a 8216 0 8217 . It is strongly recommended to not change the LBISTFREQU parameter in LBISTCTRL1 after this bit has been set to 1 because there is no guarantee that the new frequency parameter value will be transferred to the LBIST controller in time before the next LBIST run is started from user side i.e. LBISTCTRL0.LBISTREQ is set to 1 .

pub fn patterns( self ) -> RegisterField<2, 0x3ffff, 1, 0, u32, Lbistctrl0_SPEC, RW>

LBIST Pattern Number PATTERNS. This field defines the number of scan patterns i.e. scan loads which will be executed during the LBIST procedure. Please note that the value programmed to this field determines the number scan capture phases not the number of scan chain load unload phases i.e. a value of 0x00001 will result in two scan chain loads with on capture in between a value of 0x00002 will result in 3 scan chain loads with 2 captures etc. . Consequently a value of 0x00000 is not valid because no capture would be executed in this case.

pub fn lbistdone( self ) -> RegisterField<28, 0x1, 1, 0, Lbistdone, Lbistctrl0_SPEC, R>

LBIST Execution Indicator LBISTDONE. This bit indicates the actual LBIST controller execution status

pub fn lbisterrinj(self) -> RegisterFieldBool<30, 1, 0, Lbistctrl0_SPEC, RW>

LBIST Test Mode Alarm Error Injection. If written high this bit trigger both the LBIST and the test mode alarm. This is required to allow self testing of all LBIST and test mode related safety mechanisms in the TCU. The bit will be reset automatically once the LBIST and test mode alarm indicator signals from TCU are asserted. From these indicator signals SCU will also generate corresponding alarm trigger signals for SMU.

pub fn lbistreqred(self) -> RegisterFieldBool<31, 1, 0, Lbistctrl0_SPEC, W>

LBIST Request Redundancy. This bit represents the safety double of LBISTCTRL0.LBISTREQ. In order to generate a new LBIST request both LBISTREQRED and LBISTREQ bits must be set to high due to safety reasons. The request will only be approved if LBISTCTRL0.LBISTDONE bit reflects a 0 value. If read this bit always returns a 0 .