Type Definition tc37xpd::scu::Lbistctrl1

pub type Lbistctrl1 = RegValueT<Lbistctrl1_SPEC>;

Logic BIST Control 1 Register resetvalue={System Reset:0x0,CFS Value:0x54000007}

Implementations

impl Lbistctrl1

pub fn seed(self) -> RegisterField<0, 0x7ffff, 1, 0, u32, Lbistctrl1_SPEC, RW>

LBIST Seed SEED. This field determines which pattern is applied to the EDT channel inputs 1 19 during LBIST execution.

pub fn splitsh( self ) -> RegisterField<24, 0x7, 1, 0, Splitsh, Lbistctrl1_SPEC, RW>

LBIST Split Shift Selection SPLITSH. The value of this bit will allow to run LBIST with partitioned scan shift operation in order to reduce the power consumption.

pub fn body(self) -> RegisterFieldBool<27, 1, 0, Lbistctrl1_SPEC, RW>

Body Application Indicator BODY. The value of this bit will determine the static reset behavior of all GPIOs during LBIST execution. If set to low GPIOs will show a weak pull up behavior if set to high GPIOs are constrained to tri state. A high value must be written to this bit in case LBIST shall be executed for body applications.

pub fn lbistfrequ(self) -> RegisterField<28, 0xf, 1, 0, u8, Lbistctrl1_SPEC, RW>

LBIST Frequency Selection LBISTFREQU. Through this register field a pre scaler factor between 1..16 is selectable for LBIST operation clock derived from EVR oscillator . This will allow to determine the LBIST scan shift frequency. Value of these bits will be mirrored inside of LBIST controller and become effective if a new LBIST procedure has been successfully initiated from system side via LBISTCTRL0 .LBISTREQ . It is strongly recommended not to change the value of this field after LBISTCTRL0.LBISTRES has been set to high because there is no guarantee that the new frequency parameter value will be transferred to the LBIST controller in time before the next LBIST run is started from user side i.e. LBISTCTRL0.LBISTREQ is set to 1