Type Definition tc37xpd::scu::Modtrimcon0

pub type Modtrimcon0 = RegValueT<Modtrimcon0_SPEC>;

Modulation Trim 0 Configuration Register resetvalue={PowerOn Reset:0x0,After SSW execution:0x0C000000}

Implementations

impl Modtrimcon0

pub fn rnd_gain_exp( self ) -> RegisterField<0, 0x3ff, 1, 0, u16, Modtrimcon0_SPEC, RW>

XCORR Trim Value RND GAIN EXP. this bit field generates output spll rnd gain exp o .

pub fn xcorr_gain( self ) -> RegisterField<10, 0x7, 1, 0, u8, Modtrimcon0_SPEC, RW>

XCORR GAIN Trim Value XCORR GAIN. this bit field generates output spll xcorr gain o .

pub fn xcorr_target( self ) -> RegisterField<16, 0x3ff, 1, 0, u16, Modtrimcon0_SPEC, RW>

XCORR TARGET Trim Value XCORR TARGET. this bit field generates output spll xcorr target o .

pub fn hys_cal(self) -> RegisterField<26, 0xf, 1, 0, u8, Modtrimcon0_SPEC, RW>

Hysteresis Calibration Value HYS CAL. this bit field generates output spll hysteresis cal o .

pub fn ab_cal_en(self) -> RegisterFieldBool<30, 1, 0, Modtrimcon0_SPEC, RW>

Alpha Beta Calibration enable AB CAL EN. this bit field generates output spll en alpha beta cal o .

pub fn cal_en(self) -> RegisterFieldBool<31, 1, 0, Modtrimcon0_SPEC, RW>

Calibration enable CAL EN. this bit field generates output spll cal enable o .