Type Definition tc37xpd::scu::Modtrimcon1

pub type Modtrimcon1 = RegValueT<Modtrimcon1_SPEC>;

Modulation Trim 1 Configuration Register resetvalue={PowerOn Reset:0x0}

Implementations

impl Modtrimcon1

pub fn xcorr_tol(self) -> RegisterField<0, 0xff, 1, 0, u8, Modtrimcon1_SPEC, RW>

XCORR TOL Trim Value XCORR TOL. this bit field generates output spll xcorr tol o .

pub fn delta_rnd_gain( self ) -> RegisterField<8, 0xff, 1, 0, u8, Modtrimcon1_SPEC, RW>

DELTA RND GAIN Trim Value DELTA RND GAIN. this bit field generates output spll delta rnd gain o .

pub fn xcorr_avg_sel( self ) -> RegisterField<16, 0x3, 1, 0, u8, Modtrimcon1_SPEC, RW>

XCORR AVG Selection XCORR AVG SEL. this bit field generates output spll xcorr avg sel o .

pub fn xcorr_sel_delay( self ) -> RegisterField<18, 0x3, 1, 0, u8, Modtrimcon1_SPEC, RW>

XCORR SEL Delay XCORR SEL DELAY. this bit field generates output spll xcorr sel delay o .

pub fn sysemcdelay( self ) -> RegisterField<24, 0x3, 1, 0, u8, Modtrimcon1_SPEC, RW>

System PLL EMC delay enable SYSEMCDELAY

pub fn peremcdelay( self ) -> RegisterField<26, 0x3, 1, 0, u8, Modtrimcon1_SPEC, RW>

Peripheral PLL EMC delay enable PEREMCDELAY