Type Definition tc37xpd::scu::Modtrimstat

pub type Modtrimstat = RegValueT<Modtrimstat_SPEC>;

Modulation Trim Status Register resetvalue={PowerOn Reset:0x0}

Implementations

impl Modtrimstat

pub fn xcorr(self) -> RegisterField<0, 0x3ff, 1, 0, u16, Modtrimstat_SPEC, R>

XCORR Result XCORR. this bit field displays input spll xcorr i .

pub fn xcorr_valid(self) -> RegisterFieldBool<15, 1, 0, Modtrimstat_SPEC, R>

XCORR Valid Result XCORR VALID. this bit field displays input spll xcorr valid i .

pub fn xcorr2low(self) -> RegisterFieldBool<16, 1, 0, Modtrimstat_SPEC, R>

XCORR to Low Result XCORR2LOW. this bit field displays input spll xcorr2low i .

pub fn xcorr2high(self) -> RegisterFieldBool<17, 1, 0, Modtrimstat_SPEC, R>

XCORR to High Result XCORR2HIGH. this bit field displays input spll xcorr2high i .