Type Definition tc37xpd::scu::Osccon

pub type Osccon = RegValueT<Osccon_SPEC>;

OSC Control Register resetvalue={System Reset:0x258,System Reset:0x10}

Implementations

impl Osccon

pub fn plllv(self) -> RegisterField<1, 0x1, 1, 0, Plllv, Osccon_SPEC, R>

Oscillator for PLL Valid Low Status Bit PLLLV. This bit indicates if the frequency output f osc of the oscillator is above the lower threshold frequency f LV i.e. usable for the DCO part of the PLL. This is checked by the Oscillator Watchdog of the PLL using f BACK . f LV f oscref 0 96 0 31 f BACKT

pub fn oscres(self) -> RegisterField<2, 0x1, 1, 0, Oscres, Osccon_SPEC, W>

Oscillator Watchdog Reset OSCRES. Always read as zero.

pub fn gainsel(self) -> RegisterField<3, 0x3, 1, 0, Gainsel, Osccon_SPEC, RW>

Oscillator Gain Selection GAINSEL. In Normal Mode this value should not be changed from the reset value 11 B . When using Vext 3.3V the LVDS bias distributor has to be adjusted to 3.3V supply via P21 LPCR2.PS 0 otherwise the oscillator gain can be too low for a reliable oscillator startup at cold temperature. In case of using Vext 5V the LVDS bias distributor setting stays at the reset value P21 LPCR2.PS 1.

pub fn mode(self) -> RegisterField<5, 0x3, 1, 0, Mode, Osccon_SPEC, RW>

Oscillator Mode MODE. This bit field defines which mode can be used and if the oscillator entered the Power Saving Mode or not. In Test Mode the shaper can be bypassed even if the oscillator is in the Oscillator Power Saving Mode. This is controlled by the TCU.

pub fn shby(self) -> RegisterField<7, 0x1, 1, 0, Shby, Osccon_SPEC, RW>

Shaper Bypass SHBY

pub fn pllhv(self) -> RegisterField<8, 0x1, 1, 0, Pllhv, Osccon_SPEC, R>

Oscillator for PLL Valid High Status Bit PLLHV. This bit indicates if the frequency output f osc of the oscillator is below the upper threshold frequency f HV i.e. usable for the DCO part of the PLL. This is checked by the Oscillator Watchdog of the PLL using f BACK . f HV f oscref 1 04 0 29 f BACKT

pub fn hysen(self) -> RegisterField<9, 0x1, 1, 0, Hysen, Osccon_SPEC, RW>

Hysteresis Enable

pub fn hysctl(self) -> RegisterField<10, 0x3, 1, 0, Hysctl, Osccon_SPEC, RW>

Hysteresis Control

pub fn ampctl(self) -> RegisterField<12, 0x3, 1, 0, Ampctl, Osccon_SPEC, RW>

Amplitude Control

pub fn oscval(self) -> RegisterField<16, 0x1f, 1, 0, u8, Osccon_SPEC, RW>

OSC Frequency Value OSCVAL. This bit field defines the divider value that generates the reference clock that is supervised by the oscillator watchdog. f OSC 160 160 OSCCON.OSCVAL 160 160 1 160 160 16 160 MHz.

pub fn apren(self) -> RegisterField<23, 0x1, 1, 0, Apren, Osccon_SPEC, RW>

Amplitude Regulation Enable APREN. This bit field enables and disables Amplitude Regulation mode. When enabled the bit field GAINSEL limits the maximum gain.

pub fn cap0en(self) -> RegisterField<24, 0x1, 1, 0, Cap0En, Osccon_SPEC, RW>

Capacitance 0 Enable CAP0EN. Total capacitance for each XTAL1 and XTAL2 is the sum of the enabled capacitance 0 to 3.

pub fn cap1en(self) -> RegisterField<25, 0x1, 1, 0, Cap1En, Osccon_SPEC, RW>

Capacitance 1 Enable CAP1EN. Total capacitance for each XTAL1 and XTAL2 is the sum of the enabled capacitance 0 to 3.

pub fn cap2en(self) -> RegisterField<26, 0x1, 1, 0, Cap2En, Osccon_SPEC, RW>

Capacitance 2 Enable CAP2EN. Total capacitance for each XTAL1 and XTAL2 is the sum of the enabled capacitance 0 to 3.

pub fn cap3en(self) -> RegisterField<27, 0x1, 1, 0, Cap3En, Osccon_SPEC, RW>

Capacitance 3 Enable CAP3EN. Total capacitance for each XTAL1 and XTAL2 is the sum of the enabled capacitance 0 to 3.