Type Definition tc37xpd::scu::Pdrr

pub type Pdrr = RegValueT<Pdrr_SPEC>;

Pattern Detection Result Register resetvalue={Application Reset:0x0FF}

Implementations

impl Pdrr

pub fn pdr0(self) -> RegisterFieldBool<0, 1, 0, Pdrr_SPEC, R>

Pattern Detection Result of Channel 7 PDR7. This bit monitors the output status of the pattern detection for the output channel y.

pub fn pdr1(self) -> RegisterFieldBool<1, 1, 0, Pdrr_SPEC, R>

Pattern Detection Result of Channel 7 PDR7. This bit monitors the output status of the pattern detection for the output channel y.

pub fn pdr2(self) -> RegisterFieldBool<2, 1, 0, Pdrr_SPEC, R>

Pattern Detection Result of Channel 7 PDR7. This bit monitors the output status of the pattern detection for the output channel y.

pub fn pdr3(self) -> RegisterFieldBool<3, 1, 0, Pdrr_SPEC, R>

Pattern Detection Result of Channel 7 PDR7. This bit monitors the output status of the pattern detection for the output channel y.

pub fn pdr4(self) -> RegisterFieldBool<4, 1, 0, Pdrr_SPEC, R>

Pattern Detection Result of Channel 7 PDR7. This bit monitors the output status of the pattern detection for the output channel y.

pub fn pdr5(self) -> RegisterFieldBool<5, 1, 0, Pdrr_SPEC, R>

Pattern Detection Result of Channel 7 PDR7. This bit monitors the output status of the pattern detection for the output channel y.

pub fn pdr6(self) -> RegisterFieldBool<6, 1, 0, Pdrr_SPEC, R>

Pattern Detection Result of Channel 7 PDR7. This bit monitors the output status of the pattern detection for the output channel y.

pub fn pdr7(self) -> RegisterFieldBool<7, 1, 0, Pdrr_SPEC, R>

Pattern Detection Result of Channel 7 PDR7. This bit monitors the output status of the pattern detection for the output channel y.