Type Definition tc37xpd::scu::Plltrim

pub type Plltrim = RegValueT<Plltrim_SPEC>;

PLLTrim Register resetvalue={PowerOn Reset:0x080008000,CFS Value:0x080008000}

Implementations

impl Plltrim

pub fn syspllivr(self) -> RegisterField<0, 0xf, 1, 0, u8, Plltrim_SPEC, RW>

System PLL IVR Trim Value SYSPLLIVR. this bit field generates output spll ivr adj o .

pub fn sysalpha(self) -> RegisterField<4, 0x1f, 1, 0, u8, Plltrim_SPEC, RW>

System PLL Alpha Trim Value SYSALPHA. this bit field generates output spll alpha o .

pub fn sysbeta(self) -> RegisterField<9, 0x1f, 1, 0, u8, Plltrim_SPEC, RW>

System PLL Beta Trim Value SYSBETA. this bit field generates output spll beta o .

pub fn sysaben(self) -> RegisterFieldBool<14, 1, 0, Plltrim_SPEC, RW>

System PLL Alpha and Beta Trim enable SYSABEN. this bit field generates output spll set IFparams o .

pub fn sysivr_filt_dis( self ) -> RegisterField<15, 0x1, 1, 0, SysivrFiltDis, Plltrim_SPEC, RW>

System PLL IVR reference voltage filter enable. this bit field generates output spll ivr filt dis o .

pub fn perpllivr(self) -> RegisterField<16, 0xf, 1, 0, u8, Plltrim_SPEC, RW>

Peripheral PLL IVR Trim Value PERPLLIVR. this bit field generates output ppll ivr adj o .

pub fn peralpha(self) -> RegisterField<20, 0x1f, 1, 0, u8, Plltrim_SPEC, RW>

Peripheral PLL Alpha Trim Value PERALPHA. this bit field generates output ppll alpha o .

pub fn perbeta(self) -> RegisterField<25, 0x1f, 1, 0, u8, Plltrim_SPEC, RW>

Peripheral PLL Beta Trim Value PERBETA. this bit field generates output ppll beta o .

pub fn peraben(self) -> RegisterFieldBool<30, 1, 0, Plltrim_SPEC, RW>

Peripheral PLL Alpha and Beta Trim enable PERABEN. this bit field generates output ppll set IFparams o .

pub fn perivr_filt_dis( self ) -> RegisterField<31, 0x1, 1, 0, PerivrFiltDis, Plltrim_SPEC, RW>

Peripheral PLL IVR reference voltage filter disable. this bit field generates output ppll ivr filt dis o .