Type Definition tc37xpd::scu::Pmcsr1

pub type Pmcsr1 = RegValueT<Pmcsr1_SPEC>;

Power Management Control and Status Register resetvalue={Application Reset:0x100}

Implementations

impl Pmcsr1

pub fn reqslp(self) -> RegisterField<0, 0x3, 1, 0, Reqslp, Pmcsr1_SPEC, RW>

Idle Mode and Sleep Mode Request REQSLP. In Idle Mode or Sleep Mode these bits are cleared in response to an interrupt for the CPU or when bit 15 of the corresponding CPU Watchdog Timer register bit WDTCPUxSR.TIM 15 changes from 0 to 1. In Standby Mode these bits are cleared on wake up. REQSLP maybe written only when either CPU or Safety ENDINIT bits are set to 0. CPU ENDINIT bit has to be set back after REQSLP is written for the mode transition to take place. In case of Safety ENDINIT the mode transition will be issued immediately and does not wait till Safety ENDINIT is set back to 1 again. No explicit acknowledge to SCU on System Standby request. System Standby entry in a few cycles.

pub fn pmst(self) -> RegisterField<8, 0x7, 1, 0, Pmst, Pmcsr1_SPEC, R>

Power management Status PMST. This bit field reflects the current status of the CPU.