Type Definition tc37xpd::scu::Pmswcr1

pub type Pmswcr1 = RegValueT<Pmswcr1_SPEC>;

Standby and Wake up Control Register 1 resetvalue={Cold PowerOn Reset:0x1000000}

Implementations

impl Pmswcr1

pub fn iradis(self) -> RegisterField<12, 0x1, 1, 0, Iradis, Pmswcr1_SPEC, RW>

Idle Request Acknowledge Sequence Disable IRADIS. This bit enables SCU Idle Request Acknowledge sequence to all modules on Standby entry. IRADIS bit has no effect incase of Standby entry triggered via PWRWKEN register bit. This bit shall be set before Standby entry to disable Idle request acknowledge sequence so that standby request is not blocked by a pending reset idle request acknowledge sequence.

pub fn stbyeven(self) -> RegisterField<27, 0x1, 1, 0, Stbyeven, Pmswcr1_SPEC, W>

Standby Entry Event configuration enable STBYEVEN