Type Definition tc37xpd::scu::Pmtrcsr0

pub type Pmtrcsr0 = RegValueT<Pmtrcsr0_SPEC>;

Power Management Transition Control and Status Register 0 resetvalue={Cold PowerOn Reset:0x0}

Implementations

impl Pmtrcsr0

pub fn ljten(self) -> RegisterField<0, 0x1, 1, 0, Ljten, Pmtrcsr0_SPEC, RW>

Load Jump Timer Enable LJTEN. This bit field enables the usage of load jump timer.

pub fn ljtoven(self) -> RegisterField<1, 0x1, 1, 0, Ljtoven, Pmtrcsr0_SPEC, RW>

Load Jump Timer Overflow Enable LJTOVEN. This bit field enables the update of LJTOV status bit on timer overflow or time out.

pub fn ljtovien( self ) -> RegisterField<2, 0x1, 1, 0, Ljtovien, Pmtrcsr0_SPEC, RW>

Load Jump Timer Overflow Interrupt Enable LJTOVIEN. This bit field enables the activation of interrupt on timer overflow or time out.

pub fn ljtstrt(self) -> RegisterField<3, 0x1, 1, 0, Ljtstrt, Pmtrcsr0_SPEC, RW>

Load Jump Timer Start LJTSTRT. This bit field starts Load jump timer. This is intended for test purposes. The LJTSTRT remains set on a write and is cleared when LJTOV bit is set if LJTOVEN bit is enabled.

pub fn ljtstp(self) -> RegisterField<4, 0x1, 1, 0, Ljtstp, Pmtrcsr0_SPEC, RW>

Load Jump Timer Stop LJTSTP. This bit field stops Load jump timer. This is intended for test purposes. The LJTSTP remains set on a write and is to be explicitly cleared by software. The LJTSTP stops the counter at the current value and timer re starts from that value when LJTSTP is cleared and LJTSTRT is set.

pub fn ljtclr(self) -> RegisterField<5, 0x1, 1, 0, Ljtclr, Pmtrcsr0_SPEC, W>

Load Jump Timer Clear LJTCLR. This bit field clear Load jump timer count. This is intended for test purposes. This bit resets LJT and clears LJTRUN if LJTEN bit is set.

pub fn sdstep(self) -> RegisterField<12, 0xf, 1, 0, u8, Pmtrcsr0_SPEC, RW>

Droop Voltage Step vdroop step i SDSTEP. This bit field defines the voltage offset for droop compensation on a load jump to the EVRC setpoint value. The request is made via SCU PMTRCSR3 .VDROOPREQ sd droop cntr i on an anticipated load jump with a voltage offset equal to the SDSTEP x 5 160 mV. The droop step is a positive offset if VDROOPREQ 160 160 01b and is a negative offset if VDROOPREQ 160 160 10b and no offset is applied if VDROOPREQ 160 160 00b. Maximum Droop 80 160 mV.

pub fn vdten(self) -> RegisterField<16, 0x1, 1, 0, Vdten, Pmtrcsr0_SPEC, RW>

Voltage Droop Timer Enable VDTEN. This bit field enables the usage of Voltage Droop timer.

pub fn vdtoven(self) -> RegisterField<17, 0x1, 1, 0, Vdtoven, Pmtrcsr0_SPEC, RW>

Voltage Droop Timer Overflow Enable VDTOVEN. This bit field enables the update of VDTOV status bit on timer overflow or time out.

pub fn vdtovien( self ) -> RegisterField<18, 0x1, 1, 0, Vdtovien, Pmtrcsr0_SPEC, RW>

Voltage Droop Timer Overflow Interrupt Enable VDTOVIEN. This bit field enables the activation of interrupt on timer overflow or time out.

pub fn vdtstrt(self) -> RegisterField<19, 0x1, 1, 0, Vdtstrt, Pmtrcsr0_SPEC, RW>

Voltage Droop Timer Start VDTSTRT. This bit field starts Voltage Droop timer. This is intended for test purposes. The VDTSTRT remains set on a write and is cleared when VDTOV bit is set if VDTOVEN bit is enabled.

pub fn vdtstp(self) -> RegisterField<20, 0x1, 1, 0, Vdtstp, Pmtrcsr0_SPEC, RW>

Voltage Droop Timer Stop VDTSTP. This bit field stops Voltage Droop timer. SCU cancels the droop request via signal sd droop cntr i 160 160 00. This is intended for test purposes. The VDTSTP remains set on a write and is to be explicitly cleared by software. The VDTSTP stops the counter at the current value and timer re starts from that value when VDTSTP is cleared and VDTSTRT is set.

pub fn vdtclr(self) -> RegisterField<21, 0x1, 1, 0, Vdtclr, Pmtrcsr0_SPEC, W>

Voltage Droop Timer Clear VDTCLR. This bit field clear Voltage Droop timer count. This is intended for test purposes. This bit resets VDT and clears VDTRUN if VDTEN bit is set.

pub fn lpslpen(self) -> RegisterField<29, 0x1, 1, 0, Lpslpen, Pmtrcsr0_SPEC, RW>

EVRC Low Power Mode activation on a Sleep Request LPSLPEN. This bit field enables the activation of LPM EVRC mode on a sleep request. If this bit is set SCU sends lp enable signal on Sleep request to PMS which is OR ed with EVRSDCTRL2.EVRCMOD value and is provided to EVRC regulator.