Type Definition tc37xpd::scu::Pmtrcsr3

pub type Pmtrcsr3 = RegValueT<Pmtrcsr3_SPEC>;

Power Management Transition Control and Status Register 3 resetvalue={Cold PowerOn Reset:0x0}

Implementations

impl Pmtrcsr3

pub fn vdroopreq( self ) -> RegisterField<0, 0x3, 1, 0, Vdroopreq, Pmtrcsr3_SPEC, RW>

Voltage Droop Request VDROOPREQ. This bit requests a Voltage Droop consequently leading to Voltage Droop Timer start and VDTRUN bit being set if VDTEN 1 . The request is not taken if VDTRUN bit is already in set state and VDT is currently running. The request is also not taken if VDTOV bit is set AND VDTOVEN bit is enabled . The droop step is a positive offset if sd droop cntr i 160 160 01 and is a negative offset if sd droop cntr i 160 160 10 and no offset is applied if sd droop cntr i 160 160 00 and is applied immediately.

pub fn vdtov(self) -> RegisterField<8, 0x1, 1, 0, Vdtov, Pmtrcsr3_SPEC, R>

Voltage Droop Timer Overflow Status VDTOV. This status bit indicates that the Voltage Droop timer compare match has happened. if VDTOVEN bit is enabled then VDTOV can only be cleared by explicitly via VDTOVCLR bit. if VDTOVEN bit is disabled VDTOV is cleared on a taken Voltage Droop Request A new Voltage Droop request is taken only if both LJT amp VDT are not currently running and no active Voltage Droop request is being processed . VDTOV being set will lead to an interrupt if VDTOVIEN is enabled. Incase SDVOK is set by EVRC before VDT compare match VDTOV bit is not set.

pub fn vdtovclr( self ) -> RegisterField<12, 0x1, 1, 0, Vdtovclr, Pmtrcsr3_SPEC, W>

Voltage Droop Timer Overflow Status Clear VDTOVCLR. This bit clears VDTOV status bit if VDTOVEN bit is enabled. If VDTOVEN bit is disabled this bit has no effect. This bit always reads as 0.

pub fn vdtcnt(self) -> RegisterField<16, 0x3ff, 1, 0, u16, Pmtrcsr3_SPEC, R>

Voltage Droop Timer Value VDTCNT. This bit field reflects the current Voltage Droop timer value. VDTCNT value is cleared on timer overflow and on a taken Voltage Droop Request. X us is the compare value. LSB 1 us. Total range 65.5 ms