Type Definition tc37xpd::scu::Pmtrcsr4

pub type Pmtrcsr4 = RegValueT<Pmtrcsr4_SPEC>;

Power Management Transition Control and Status Register 4 resetvalue={Cold PowerOn Reset:0x0}

Implementations

impl Pmtrcsr4

pub fn rstdroopen(self) -> RegisterFieldBool<0, 1, 0, Pmtrcsr4_SPEC, RW>

Warm Reset Droop Enable RSTDROOPEN. This bit field enables negative voltage droop incase of a warm reset request. A negative droop step is requested via sd droop cntr i 10 with Reset droop value as configured in SDSTEP register. By default Reset Droop is disabled.

pub fn bprstdroopen(self) -> RegisterFieldBool<1, 1, 0, Pmtrcsr4_SPEC, W>

Bit Protection RSTDROOPEN BPRSTDROOPEN. Setting this bit enables that bit RSTDROOPEN can be changed in this write operation.

pub fn rstvdroop( self ) -> RegisterField<2, 0x3, 1, 0, Rstvdroop, Pmtrcsr4_SPEC, RW>

Voltage Droop during Reset RSTVDROOP

pub fn rstsdstep(self) -> RegisterField<12, 0xf, 1, 0, u8, Pmtrcsr4_SPEC, RW>

Droop Voltage Step vdroop step i RSTSDSTEP. This bit field RSTSDSTEP defines the voltage offset for droop compensation on a reset jump to the EVRC setpoint value. The request is made via sd droop cntr i signal on an anticipated load jump with a voltage offset equal to the RSTSDSTEP x 5 160 mV. Maximum Droop 80 160 mV.