Type Definition tc37xpd::scu::Rstcon2

pub type Rstcon2 = RegValueT<Rstcon2_SPEC>;

Additional Reset Control Register resetvalue={Cold PowerOn Reset:0x0,Cold PowerOn Reset:0x0}

Implementations

impl Rstcon2

pub fn frto(self) -> RegisterField<0, 0x1, 1, 0, Frto, Rstcon2_SPEC, RW>

Force Reset Timeout FRTO

pub fn clrc(self) -> RegisterField<1, 0x1, 1, 0, Clrc, Rstcon2_SPEC, W>

Clear Cold Reset Status CLRC. This bit simultaneously clears the sticky status bits which may indicate any previous cold reset i.e. RSTSTAT.STBYR RSTSTAT.SWD RSTSTAT.EVR33 RSTSTAT.EVRC RSTSTAT.PORST RSTSTAT.LBPORST and RSTSTAT.LBTERM .

pub fn tout20(self) -> RegisterField<2, 0x1, 1, 0, Tout20, Rstcon2_SPEC, R>

Reset Timeout Counter 20us TOUT20. This bit indicates elapsed time after the last warm reset request. This may used by a shutdown routine to schedule its activities independently from current clock settings

pub fn tout40(self) -> RegisterField<3, 0x1, 1, 0, Tout40, Rstcon2_SPEC, R>

Reset Timeout Counter 40us TOUT40. This bit indicates elapsed time after the last warm reset request. This may used by a shutdown routine to schedule its activities independently from current clock settings

pub fn tout60(self) -> RegisterField<4, 0x1, 1, 0, Tout60, Rstcon2_SPEC, R>

Reset Timeout Counter 60us TOUT60. This bit indicates elapsed time after the last warm reset request. This may used by a shutdown routine to schedule its activities independently from current clock settings

pub fn tout80(self) -> RegisterField<5, 0x1, 1, 0, Tout80, Rstcon2_SPEC, R>

Reset Timeout Counter 80us TOUT80. This bit indicates elapsed time after the last warm reset request. This may used by a shutdown routine to schedule its activities independently from current clock settings

pub fn flss(self) -> RegisterField<6, 0x1, 1, 0, Flss, Rstcon2_SPEC, R>

Flash Shutdown State Reached FLSS. Represents the state of the FSI before the last warm reset If this bit is zero after a System Reset or higher then it is possible that Flash reliability could have been affected by the reset.

pub fn cssx(self) -> RegisterField<7, 0x3f, 1, 0, CsSx, Rstcon2_SPEC, R>

CPU x Shutdown State Reached CSSx. The state of CPU x before the last warm reset If any bit is zero after an Application Reset or higher then it is possible that SRAM content could have been corrupted by the reset For products with fewer CPUs only the LSBs are active and unused upper bits will always read 8216 1 8217

pub fn mcdsss(self) -> RegisterField<15, 0x1, 1, 0, Mcdsss, Rstcon2_SPEC, R>

MCDS Shutdown State Reached MCDSSS. This bit is valid only in an Emulation Device. This bit represents the state of MCDS or MiniMCDS before the last warm reset If this bit is zero after a warm PORST then it is possible that trace content could have been corrupted by the reset

pub fn usrinfo(self) -> RegisterField<16, 0xffff, 1, 0, u16, Rstcon2_SPEC, RW>

User Information USRINFO. User data register Cleared only on Cold Power on reset . This may be used by an application to store information which must survive all warm resets