Type Definition tc37xpd::scu::Rststat

pub type Rststat = RegValueT<Rststat_SPEC>;

Reset Status Register resetvalue={Cold PowerOn Reset:0x13810000,Cold PowerOn Reset:0x13810000}

Implementations

impl Rststat

pub fn esr0(self) -> RegisterField<0, 0x1, 1, 0, Esr0, Rststat_SPEC, R>

Reset Request Trigger Reset Status for ESR0 ESR0

pub fn esr1(self) -> RegisterField<1, 0x1, 1, 0, Esr1, Rststat_SPEC, R>

Reset Request Trigger Reset Status for ESR1 ESR1

pub fn smu(self) -> RegisterField<3, 0x1, 1, 0, Smu, Rststat_SPEC, R>

Reset Request Trigger Reset Status for SMU SMU. See SMU section for SMU trigger sources including Watchdog Timers

pub fn sw(self) -> RegisterField<4, 0x1, 1, 0, Sw, Rststat_SPEC, R>

Reset Request Trigger Reset Status for SW SW

pub fn stm0(self) -> RegisterField<5, 0x1, 1, 0, Stm0, Rststat_SPEC, R>

Reset Request Trigger Reset Status for STM0 Compare Match STM0

pub fn stm1(self) -> RegisterField<6, 0x1, 1, 0, Stm1, Rststat_SPEC, R>

Reset Request Trigger Reset Status for STM1 Compare Match If Product has STM1 STM1

pub fn stm2(self) -> RegisterField<7, 0x1, 1, 0, Stm2, Rststat_SPEC, R>

Reset Request Trigger Reset Status for STM2 Compare Match If Product has STM2 STM2

pub fn porst(self) -> RegisterField<16, 0x1, 1, 0, Porst, Rststat_SPEC, R>

Reset Request Trigger Reset Status for PORST PORST. This bit is also set if the bits CB0 CB1 and CB3 are set in parallel.

pub fn cb0(self) -> RegisterField<18, 0x1, 1, 0, Cb0, Rststat_SPEC, R>

Reset Request Trigger Reset Status for Cerberus System Reset CB0

pub fn cb1(self) -> RegisterField<19, 0x1, 1, 0, Cb1, Rststat_SPEC, R>

Reset Request Trigger Reset Status for Cerberus Debug Reset CB1

pub fn cb3(self) -> RegisterField<20, 0x1, 1, 0, Cb3, Rststat_SPEC, R>

Reset Request Trigger Reset Status for Cerberus Application Reset CB3

pub fn tp(self) -> RegisterField<21, 0x1, 1, 0, Tp, Rststat_SPEC, R>

Reset Request Trigger Reset Status for Tuning Protection TP

pub fn tcu(self) -> RegisterField<22, 0x1, 1, 0, Tcu, Rststat_SPEC, R>

Reset Request Trigger Reset Status for TCU TCU

pub fn evrc(self) -> RegisterField<23, 0x1, 1, 0, Evrc, Rststat_SPEC, R>

Reset Request Trigger Reset Status for EVRC EVRC

pub fn evr33(self) -> RegisterField<24, 0x1, 1, 0, Evr33, Rststat_SPEC, R>

Reset Request Trigger Reset Status for EVR33 EVR33

pub fn swd(self) -> RegisterField<25, 0x1, 1, 0, Swd, Rststat_SPEC, R>

Reset Request Trigger Reset Status for Supply Watchdog SWD SWD. The Supply Watchdog trigger is described in Power Management Controller 8220 Supply Monitoring 8221 chapter

pub fn hsms(self) -> RegisterField<26, 0x1, 1, 0, Hsms, Rststat_SPEC, R>

Reset Request Trigger Reset Status for HSM System Reset HSM S HSMS

pub fn hsma(self) -> RegisterField<27, 0x1, 1, 0, Hsma, Rststat_SPEC, R>

Reset Request Trigger Reset Status for HSM Application Reset HSM A HSMA

pub fn stbyr(self) -> RegisterField<28, 0x1, 1, 0, Stbyr, Rststat_SPEC, R>

Reset Request Trigger Reset Status for Standby Regulator Watchdog STBYR STBYR

pub fn lbporst(self) -> RegisterField<29, 0x1, 1, 0, Lbporst, Rststat_SPEC, R>

LBIST termination due to PORST. This bitfield indicates if the LBIST was early terminated due to the occurrence of a Power On Reset. If the status of this bitfield is 0 the application must still check the LBTERM to check if the LBIST was terminated properly. This bitfield is cleared when the RSTCON2.CLRC is set.

pub fn lbterm(self) -> RegisterField<30, 0x1, 1, 0, Lbterm, Rststat_SPEC, R>

LBIST was properly terminated. This bitfield indicates if the LBIST was terminated properly. This bitfield is cleared when the RSTCON2.CLRC is set.