Type Definition tc37xpd::scu::Rtid

pub type Rtid = RegValueT<Rtid_SPEC>;

Redesign Tracing Identification Register resetvalue={System Reset:0x0}

Implementations

impl Rtid

pub fn rt0(self) -> RegisterField<0, 0x1, 1, 0, Rt0, Rtid_SPEC, R>

Redesign Trace Bit 15 RT15. RTx can be used e.g. for minor redesign stepping identification purposes or enabling of features. The assignment of individual bits in this field will be defined in the TLDS and DfT Specification. Bit 15 Allow access to FSI in test mode Bit 14 Allow STP to be disabled by TCU Bit 13 Unlock Cerberus ocds hw lock

pub fn rt1(self) -> RegisterField<1, 0x1, 1, 0, Rt1, Rtid_SPEC, R>

Redesign Trace Bit 15 RT15. RTx can be used e.g. for minor redesign stepping identification purposes or enabling of features. The assignment of individual bits in this field will be defined in the TLDS and DfT Specification. Bit 15 Allow access to FSI in test mode Bit 14 Allow STP to be disabled by TCU Bit 13 Unlock Cerberus ocds hw lock

pub fn rt2(self) -> RegisterField<2, 0x1, 1, 0, Rt2, Rtid_SPEC, R>

Redesign Trace Bit 15 RT15. RTx can be used e.g. for minor redesign stepping identification purposes or enabling of features. The assignment of individual bits in this field will be defined in the TLDS and DfT Specification. Bit 15 Allow access to FSI in test mode Bit 14 Allow STP to be disabled by TCU Bit 13 Unlock Cerberus ocds hw lock

pub fn rt3(self) -> RegisterField<3, 0x1, 1, 0, Rt3, Rtid_SPEC, R>

Redesign Trace Bit 15 RT15. RTx can be used e.g. for minor redesign stepping identification purposes or enabling of features. The assignment of individual bits in this field will be defined in the TLDS and DfT Specification. Bit 15 Allow access to FSI in test mode Bit 14 Allow STP to be disabled by TCU Bit 13 Unlock Cerberus ocds hw lock

pub fn rt4(self) -> RegisterField<4, 0x1, 1, 0, Rt4, Rtid_SPEC, R>

Redesign Trace Bit 15 RT15. RTx can be used e.g. for minor redesign stepping identification purposes or enabling of features. The assignment of individual bits in this field will be defined in the TLDS and DfT Specification. Bit 15 Allow access to FSI in test mode Bit 14 Allow STP to be disabled by TCU Bit 13 Unlock Cerberus ocds hw lock

pub fn rt5(self) -> RegisterField<5, 0x1, 1, 0, Rt5, Rtid_SPEC, R>

Redesign Trace Bit 15 RT15. RTx can be used e.g. for minor redesign stepping identification purposes or enabling of features. The assignment of individual bits in this field will be defined in the TLDS and DfT Specification. Bit 15 Allow access to FSI in test mode Bit 14 Allow STP to be disabled by TCU Bit 13 Unlock Cerberus ocds hw lock

pub fn rt6(self) -> RegisterField<6, 0x1, 1, 0, Rt6, Rtid_SPEC, R>

Redesign Trace Bit 15 RT15. RTx can be used e.g. for minor redesign stepping identification purposes or enabling of features. The assignment of individual bits in this field will be defined in the TLDS and DfT Specification. Bit 15 Allow access to FSI in test mode Bit 14 Allow STP to be disabled by TCU Bit 13 Unlock Cerberus ocds hw lock

pub fn rt7(self) -> RegisterField<7, 0x1, 1, 0, Rt7, Rtid_SPEC, R>

Redesign Trace Bit 15 RT15. RTx can be used e.g. for minor redesign stepping identification purposes or enabling of features. The assignment of individual bits in this field will be defined in the TLDS and DfT Specification. Bit 15 Allow access to FSI in test mode Bit 14 Allow STP to be disabled by TCU Bit 13 Unlock Cerberus ocds hw lock

pub fn rt8(self) -> RegisterField<8, 0x1, 1, 0, Rt8, Rtid_SPEC, R>

Redesign Trace Bit 15 RT15. RTx can be used e.g. for minor redesign stepping identification purposes or enabling of features. The assignment of individual bits in this field will be defined in the TLDS and DfT Specification. Bit 15 Allow access to FSI in test mode Bit 14 Allow STP to be disabled by TCU Bit 13 Unlock Cerberus ocds hw lock

pub fn rt9(self) -> RegisterField<9, 0x1, 1, 0, Rt9, Rtid_SPEC, R>

Redesign Trace Bit 15 RT15. RTx can be used e.g. for minor redesign stepping identification purposes or enabling of features. The assignment of individual bits in this field will be defined in the TLDS and DfT Specification. Bit 15 Allow access to FSI in test mode Bit 14 Allow STP to be disabled by TCU Bit 13 Unlock Cerberus ocds hw lock

pub fn rt10(self) -> RegisterField<10, 0x1, 1, 0, Rt10, Rtid_SPEC, R>

Redesign Trace Bit 15 RT15. RTx can be used e.g. for minor redesign stepping identification purposes or enabling of features. The assignment of individual bits in this field will be defined in the TLDS and DfT Specification. Bit 15 Allow access to FSI in test mode Bit 14 Allow STP to be disabled by TCU Bit 13 Unlock Cerberus ocds hw lock

pub fn rt11(self) -> RegisterField<11, 0x1, 1, 0, Rt11, Rtid_SPEC, R>

Redesign Trace Bit 15 RT15. RTx can be used e.g. for minor redesign stepping identification purposes or enabling of features. The assignment of individual bits in this field will be defined in the TLDS and DfT Specification. Bit 15 Allow access to FSI in test mode Bit 14 Allow STP to be disabled by TCU Bit 13 Unlock Cerberus ocds hw lock

pub fn rt12(self) -> RegisterField<12, 0x1, 1, 0, Rt12, Rtid_SPEC, R>

Redesign Trace Bit 15 RT15. RTx can be used e.g. for minor redesign stepping identification purposes or enabling of features. The assignment of individual bits in this field will be defined in the TLDS and DfT Specification. Bit 15 Allow access to FSI in test mode Bit 14 Allow STP to be disabled by TCU Bit 13 Unlock Cerberus ocds hw lock

pub fn rt13(self) -> RegisterField<13, 0x1, 1, 0, Rt13, Rtid_SPEC, R>

Redesign Trace Bit 15 RT15. RTx can be used e.g. for minor redesign stepping identification purposes or enabling of features. The assignment of individual bits in this field will be defined in the TLDS and DfT Specification. Bit 15 Allow access to FSI in test mode Bit 14 Allow STP to be disabled by TCU Bit 13 Unlock Cerberus ocds hw lock

pub fn rt14(self) -> RegisterField<14, 0x1, 1, 0, Rt14, Rtid_SPEC, R>

Redesign Trace Bit 15 RT15. RTx can be used e.g. for minor redesign stepping identification purposes or enabling of features. The assignment of individual bits in this field will be defined in the TLDS and DfT Specification. Bit 15 Allow access to FSI in test mode Bit 14 Allow STP to be disabled by TCU Bit 13 Unlock Cerberus ocds hw lock

pub fn rt15(self) -> RegisterField<15, 0x1, 1, 0, Rt15, Rtid_SPEC, R>

Redesign Trace Bit 15 RT15. RTx can be used e.g. for minor redesign stepping identification purposes or enabling of features. The assignment of individual bits in this field will be defined in the TLDS and DfT Specification. Bit 15 Allow access to FSI in test mode Bit 14 Allow STP to be disabled by TCU Bit 13 Unlock Cerberus ocds hw lock