Type Definition tc37xpd::scu::Seisr

pub type Seisr = RegValueT<Seisr_SPEC>;

Safety ENDINIT Timeout Status Register resetvalue={Application Reset:0x0FFFC0000}

Implementations

impl Seisr

pub fn ae(self) -> RegisterField<0, 0x1, 1, 0, Ae, Seisr_SPEC, R>

SEICON0 Access Error Status Flag AE. This bit is set when an illegal Password Access or Modify Access to register SEICON0 was attempted. This bit is only cleared on a valid SEICON0.ENDINIT Modify Access

pub fn oe(self) -> RegisterField<1, 0x1, 1, 0, Oe, Seisr_SPEC, R>

SEI Timeout Overflow Error Status Flag OE. This bit is set when SEISR.TIM overflows from FFFF H to FFFC H . This bit is only cleared on a valid SEICON0 Modify Access.

pub fn is0(self) -> RegisterField<2, 0x1, 1, 0, Is0, Seisr_SPEC, R>

SEI Timeout Input Clock Status IS0 IS1. Bit IS0 and IS1 should be programmed together. These bits inidicate the current Safety ENDINIT Timeout Counter clock period. They are updated with the state of bits SEICON1.IR0 and SEICON1.IR1 after a valid SEICON0 Modify Access.

pub fn ds(self) -> RegisterField<3, 0x1, 1, 0, Ds, Seisr_SPEC, R>

SEI Enable Disable Status Flag DS

pub fn to(self) -> RegisterField<4, 0x1, 1, 0, To, Seisr_SPEC, R>

SEI Time Out Mode Flag TO

pub fn is1(self) -> RegisterField<5, 0x1, 1, 0, Is1, Seisr_SPEC, R>

SEI Timeout Input Clock Status IS0 IS1. Bit IS0 and IS1 should be programmed together. These bits inidicate the current Safety ENDINIT Timeout Counter clock period . They are updated with the state of bits SEICON1.IR0 and SEICON1.IR1 after a valid SEICON0 Modify Access.

pub fn tim(self) -> RegisterField<16, 0xffff, 1, 0, u16, Seisr_SPEC, R>

Timer Value TIM. Reflects the current content of the Safety EINDINIT Timeout Counter. Only bits 17 and 16 are implemented in SEISR. Others return 8216 1 8217