Type Definition tc37xpd::scu::Ststat

pub type Ststat = RegValueT<Ststat_SPEC>;

Start up Status Register resetvalue={PowerOn Reset:0x08000}

Implementations

impl Ststat

pub fn hwcfg(self) -> RegisterField<0, 0xff, 1, 0, u8, Ststat_SPEC, R>

Hardware Configuration Setting HWCFG. This bit field contains the value that is used by the boot software. This bit field is updated in case of an Application Reset with the content by register SWRSTCON.SWCFG if bit SWRSTCON.SWBOOT AND RSTSTAT.SW are set. This bit field is updated in case of an Application Reset with the content of the latches of pins P14.2 P14.6 P10.5 P10.6 if bit SWRSTCON.SWBOOT OR RSTSTAT.SW are cleared and bit STSTAT.LUDIS is cleared. This bit field is left unchanged in case of an Application Reset and is not updated with the content of the latches of pins P14.2 P14.6 P10.5 P10.6 if bit SWRSTCON.SWBOOT OR RSTSTAT.SW are cleared and bit STSTAT.LUDIS is set. The observed reset value after boot depends upon the state of the HWCFG pins

pub fn ftm(self) -> RegisterField<8, 0x7f, 1, 0, u8, Ststat_SPEC, R>

Firmware Test Setting FTM. This bit field contains the value that is used by the boot software in Test Mode. This Bit field is updated in Test Mode by the TCU. In Normal Mode this bit field is updated with 0000000 B and should be ignored by the boot software.

pub fn mode(self) -> RegisterField<15, 0x1, 1, 0, Mode, Ststat_SPEC, R>

MODE MODE. This bit indicates if the Test Mode is entered or not.

pub fn fcbae(self) -> RegisterField<16, 0x1, 1, 0, Fcbae, Ststat_SPEC, R>

Flash Config. Sector Access Enable FCBAE. This bit can be cleared by setting bit STCON.CFCBAE. This bit can be set by setting bit STCON.SFCBAE. Reset value of this bit is 0

pub fn ludis(self) -> RegisterField<17, 0x1, 1, 0, Ludis, Ststat_SPEC, R>

Latch Update Disable LUDIS. This bit can be set by setting bit SYSCON.SETLUDIS. Reset value of this bit is 0

pub fn farfuse(self) -> RegisterField<18, 0x1, 1, 0, Farfuse, Ststat_SPEC, R>

FAR Fuse State FARFUSE. All FAR fuses must be blown to enter FAR state See DfT Spec Reset value of this bit is 0

pub fn trstl(self) -> RegisterFieldBool<19, 1, 0, Ststat_SPEC, R>

TRSTL Status TRSTL. This bit simply displays the value of TRSTL.

pub fn spden(self) -> RegisterField<20, 0x1, 1, 0, Spden, Ststat_SPEC, R>

Single Pin DAP Mode Enable SPDEN

pub fn otpoff(self) -> RegisterField<21, 0x1, 1, 0, Otpoff, Ststat_SPEC, R>

OTP Over ruled OTPOFF

pub fn fsioff(self) -> RegisterField<22, 0x1, 1, 0, Fsioff, Ststat_SPEC, R>

FSI Disabled FSIOFF. fsi clock disable o signal from TCU

pub fn stpovreq(self) -> RegisterField<23, 0x1, 1, 0, Stpovreq, Ststat_SPEC, R>

Start Up Protection Over ruling Request STPOVREQ

pub fn ramint(self) -> RegisterField<24, 0x1, 1, 0, Ramint, Ststat_SPEC, R>

RAM Content Security Integrity RAMINT. In normal operation this bit can be set or cleared by the application via SYSCON.RAMINTM . If a test boot mode is entered the bit is automatically cleared and cannot be set again in test mode because the content may have been altered This bit is reset only by a cold power on reset.

pub fn subfuse(self) -> RegisterField<28, 0xf, 1, 0, Subfuse, Ststat_SPEC, R>

Individual FAR Fuse State SUBFUSE. Fuse state latched after last power on reset For each bit x in this field