Type Definition tc37xpd::scu::Syspllcon0

pub type Syspllcon0 = RegValueT<Syspllcon0_SPEC>;

System PLL Configuration 0 Register resetvalue={System Reset:0x40003A00}

Implementations

impl Syspllcon0

pub fn moden(self) -> RegisterField<2, 0x1, 1, 0, Moden, Syspllcon0_SPEC, RW>

Modulation Enable MODEN. This bit controls the activation of the frequency modulation of the System PLL.

pub fn ndiv(self) -> RegisterField<9, 0x7f, 1, 0, u8, Syspllcon0_SPEC, RW>

N Divider Value NDIV. The value the N Divider operates with is NDIV 1.

pub fn pllpwd(self) -> RegisterField<16, 0x1, 1, 0, Pllpwd, Syspllcon0_SPEC, RW>

System PLL Power Saving Mode PLLPWD. If the PLL has been powered down and is getting re enabled via PLLPWD 1 a wait period of 1 ms has to be applied until it is stable without jitter.

pub fn resld(self) -> RegisterFieldBool<18, 1, 0, Syspllcon0_SPEC, W>

Restart DCO Lock Detection RESLD. Setting this bit will clear bit SYSPLLSTAT.LOCK and restart the DCO lock detection. Reading this bit returns always a zero.

pub fn pdiv(self) -> RegisterField<24, 0x7, 1, 0, u8, Syspllcon0_SPEC, RW>

P Divider Value PDIV. The value the P Divider operates with is PDIV 1.

pub fn insel(self) -> RegisterField<30, 0x3, 1, 0, Insel, Syspllcon0_SPEC, RW>

Input Selection INSEL. This bit field defines as clock source for the two PLLs System PLL and Peripheral PLL .