Type Definition tc37xpd::scu::wdts::Wdtscon1

pub type Wdtscon1 = RegValueT<Wdtscon1_SPEC>;

Safety WDT Control Register 1 resetvalue={Application Reset:0x0}

Implementations

impl Wdtscon1

pub fn clrirf(self) -> RegisterField<0, 0x1, 1, 0, Clrirf, Wdtscon1_SPEC, RW>

Clear Internal Reset Flag CLRIRF. This bit is used to request a clear of the internal flag which indicates whether a previous SMU reset has already been requested After modification the internal flag is only cleared when Safety Endinit SE is re asserted. As long as Safety ENDINIT SE is not asserted the internal flag is unchanged and continues to determine the response to a further SMU reset request. When Safety ENDINIT is reasserted the internal flag is cleared together with this bit.

pub fn ir0(self) -> RegisterField<2, 0x1, 1, 0, Ir0, Wdtscon1_SPEC, RW>

Input Frequency Request Control IR1 IR0. Bit IR0 and IR1 should be programmed together to determine the WDTx timer frequency. WDTxSR.IS0 and WDTxSR.IS1 are updated by these bits only when Safety ENDINIT SE is re asserted. As long as Safety ENDINIT is de asserted WDTxSR.IS0 and WDTxSR.IS1 control the current input frequency of the Safety Watchdog Timer. When Safety ENDINIT is reasserted WDTxSR.IS0 and WDTxSR.IS1 are updated with the values of IR0 and IR1.

pub fn dr(self) -> RegisterField<3, 0x1, 1, 0, Dr, Wdtscon1_SPEC, RW>

Disable Request Control Bit DR. This bit can only be modified when Safety ENDINIT SE is de asserted. WDTxSR.DS is updated when Safety ENDINIT is re asserted. As long as Safety ENDINIT is de asserted bit WDTxSR.DS controls the current enable disable status of the WDTx. When Safety ENDINIT is reasserted WDTxSR.DS is updated with the state of DR.

pub fn ir1(self) -> RegisterField<5, 0x1, 1, 0, Ir1, Wdtscon1_SPEC, RW>

Input Frequency Request Control IR1 IR0. Bit IR0 and IR1 should be programmed together to determine the WDTx timer frequency WDTxSR.IS0 and WDTxSR.IS1 are updated by these bits only when Safety ENDINIT SE is re asserted. As long as Safety ENDINIT is de asserted WDTxSR.IS0 and WDTxSR.IS1 control the current input frequency of the Safety Watchdog Timer. When Safety ENDINIT is reasserted WDTxSR.IS0 and WDTxSR.IS1 are updated with the values of IR0 and IR1.

pub fn ur(self) -> RegisterField<6, 0x1, 1, 0, Ur, Wdtscon1_SPEC, RW>

Unlock Restriction Request Control Bit UR. This bit can only be modified when Safety ENDINIT SE is de asserted. WDTxSR.US is updated when Safety ENDINIT is reasserted. As long as the Safety ENDINIT is cleared bit WDTxSR.US controls whether unlocking is possible at all times or only when the SMU is not in the FAULT state. When Safety ENDINIT is reasserted WDTxSR.US is updated with the state of UR.

pub fn par(self) -> RegisterField<7, 0x1, 1, 0, Par, Wdtscon1_SPEC, RW>

Password Auto sequence Request Bit PAR. This bit can only be modified when Safety ENDINIT is de asserted. WDTxSR.PAS is updated when Safety ENDINIT is reasserted. As long as Safety ENDINIT is de asserted bit WDTxSR.PAS controls password sequencing. When Safety ENDINIT is reasserted WDTxSR.PAS is updated with the state of PAR.

pub fn tcr(self) -> RegisterField<8, 0x1, 1, 0, Tcr, Wdtscon1_SPEC, RW>

Counter Check Request Bit TCR. This bit can only be modified when Safety ENDINIT SE is de asserted. WDTxSR.TCS is updated when Safety ENDINIT is re asserted. As long as Safety ENDINIT is de asserted bit WDTxSR.TCS controls whether counter check is enabled. When Safety ENDINIT is reasserted WDTxSR.TCS is updated with the state of TCR

pub fn tctr(self) -> RegisterField<9, 0x7f, 1, 0, u8, Wdtscon1_SPEC, RW>

Timer Check Tolerance Request TCTR. This bit can only be modified when Safety ENDINIT is de asserted. WDTxSR.TCT is updated when Safety ENDINIT is reasserted. As long as Safety ENDINIT is de asserted bit WDTxSR.TCT controls the tolerance of timer checks. When Safety ENDINIT is re asserted WDTxSR.TCT is updated with the state of TCTR.