Type Definition tc37xpd::scu::wdts::Wdtssr

pub type Wdtssr = RegValueT<Wdtssr_SPEC>;

Safety WDT Status Register resetvalue={Application Reset:0x0FFFC0010}

Implementations

impl Wdtssr

pub fn ae(self) -> RegisterField<0, 0x1, 1, 0, Ae, Wdtssr_SPEC, R>

Watchdog Access Error Status Flag AE. This bit is set when an illegal Password Access or Modify Access to register WDTxCON0 was attempted. This bit is only cleared when WDTxCON0.ENDINIT is set during a valid Modify Access

pub fn oe(self) -> RegisterField<1, 0x1, 1, 0, Oe, Wdtssr_SPEC, R>

Watchdog Overflow Error Status Flag OE. This bit is set when the WDTx overflows from FFFF H to 0000 H . This bit is only cleared when WDTxCON0.ENDINIT is set to 1 during a valid Modify Access.

pub fn is0(self) -> RegisterField<2, 0x1, 1, 0, Is0, Wdtssr_SPEC, R>

Watchdog Input Clock Status IS1 IS0. Bit IS0 and IS1 should be programmed together. These bits indicate the current WDTx clock rate. These bits are updated with the state of bits WDTxCON1.IR0 and WDTxCON1.IR1 after WDTxCON0.ENDINIT is written with 1 during a valid Modify Access to register WDTxCON0.

pub fn ds(self) -> RegisterField<3, 0x1, 1, 0, Ds, Wdtssr_SPEC, R>

Watchdog Enable Disable Status Flag DS. This bit is updated with the state of bit WDTxCON1.DR after WDTxCON0.ENDINIT is set during a Valid Modify Access to register WDTxCON0 and it is cleared when Time Out mode is entered.

pub fn to(self) -> RegisterField<4, 0x1, 1, 0, To, Wdtssr_SPEC, R>

Watchdog Time Out Mode Flag TO. This bit is set when Time Out Mode is entered. It is automatically cleared when Time Out Mode is left.

pub fn is1(self) -> RegisterField<5, 0x1, 1, 0, Is1, Wdtssr_SPEC, R>

Watchdog Input Clock Status IS1 IS0. Bit IS0 and IS1 should be programmed together. These bits indicate the current WDTx clock rate. These bits are updated with the state of bits WDTxCON1.IR0 and WDTxCON1.IR1 after WDTxCON0.ENDINIT is written with 1 during a valid Modify Access to register WDTxCON0.

pub fn us(self) -> RegisterField<6, 0x1, 1, 0, Us, Wdtssr_SPEC, R>

SMU Unlock Restriction Status Flag US. WDTxCON0.LCK will not be unlocked by a valid Password Access if this bit is 8216 1 8217 and the SMU is not in the FAULT state

pub fn pas(self) -> RegisterField<7, 0x1, 1, 0, Pas, Wdtssr_SPEC, R>

Password Auto sequence Status Flag PAS. This bit is updated with the state of bit WDTxCON1.PAR after WDTxCON0.ENDINIT is written with 1 during a valid Modify Access to register WDTxCON0.

pub fn tcs(self) -> RegisterField<8, 0x1, 1, 0, Tcs, Wdtssr_SPEC, R>

Timer Check Status Flag TCS. This bit is updated with the state of bit WDTxCON1.TCR after WDTxCON0.ENDINIT is written with 1 during a Valid Modify Access to register WDTxCON0.

pub fn tct(self) -> RegisterField<9, 0x7f, 1, 0, u8, Wdtssr_SPEC, R>

Timer Check Tolerance TCT. This field determines the tolerance of the timer check during Password or Check Access See TCS .This bit is updated with the state of bit WDTxCON1.TCTR after WDTxCON0.ENDINIT is written with 1 during a Valid Modify Access to register WDTxCON0.

pub fn tim(self) -> RegisterField<16, 0xffff, 1, 0, u16, Wdtssr_SPEC, R>

Timer Value TIM. Reflects the current content of the WDTx.