Type Definition tc37xpd::stm0::Icr

pub type Icr = RegValueT<Icr_SPEC>;

Interrupt Control Register resetvalue={Application Reset:0x0}

Implementations

impl Icr

pub fn cmp0en(self) -> RegisterField<0, 0x1, 1, 0, Cmp0En, Icr_SPEC, RW>

Compare Register CMP0 Interrupt Enable Control CMP0EN. This bit enables the compare match interrupt with compare register CMP0.

pub fn cmp0ir(self) -> RegisterField<1, 0x1, 1, 0, Cmp0Ir, Icr_SPEC, R>

Compare Register CMP0 Interrupt Request Flag CMP0IR. This bit indicates whether or not a compare match event of compare register CMP0 has occured. CMP0IR must be cleared by software. CMP0IR must be cleared by software and can be set by software too see ISCR register . After a STM reset operation CMP0IR is immediately set as a result of a compare match event with the reset values of the STM and the compare registers CMP0.

pub fn cmp0os(self) -> RegisterField<2, 0x1, 1, 0, Cmp0Os, Icr_SPEC, RW>

Compare Register CMP0 Interrupt Output Selection CMP0OS. This bit determines the interrupt output that is activated on a compare match event of compare register CMP0.

pub fn cmp1en(self) -> RegisterField<4, 0x1, 1, 0, Cmp1En, Icr_SPEC, RW>

Compare Register CMP1 Interrupt Enable Control CMP1EN. This bit enables the compare match interrupt with compare register CMP1.

pub fn cmp1ir(self) -> RegisterField<5, 0x1, 1, 0, Cmp1Ir, Icr_SPEC, R>

Compare Register CMP1 Interrupt Request Flag CMP1IR. This bit indicates whether or not a compare match event of compare register CMP1 has occured. CMP1IR must be cleared by software. CMP1IR must be cleared by software and can be set by software too see ISCR register . After a STM reset CMP1IR is immediately set as a result of a compare match event with the reset values of the STM and the compare register CMP1.

pub fn cmp1os(self) -> RegisterField<6, 0x1, 1, 0, Cmp1Os, Icr_SPEC, RW>

Compare Register CMP1 Interrupt Output Selection CMP1OS. This bit determines the interrupt output that is activated on a compare match event of compare register CMP1.