Type Definition tc37xpd::stm2::Cmcon

pub type Cmcon = RegValueT<Cmcon_SPEC>;

Compare Match Control Register resetvalue={Application Reset:0x0}

Implementations

impl Cmcon

pub fn msize0(self) -> RegisterField<0, 0x1f, 1, 0, Msize0, Cmcon_SPEC, RW>

Compare Register Size for CMP0 MSIZE0. This bit field determines the number of bits in register CMP0 starting from bit 0 that are used for the compare operation with the System Timer. …

pub fn mstart0(self) -> RegisterField<8, 0x1f, 1, 0, Mstart0, Cmcon_SPEC, RW>

Start Bit Location for CMP0 MSTART0. This bit field determines the lowest bit number of the 64 bit STM that is compared with the content of register CMP0 bit 0. The number of bits to be compared is defined by bit field MSIZE0. …

pub fn msize1(self) -> RegisterField<16, 0x1f, 1, 0, Msize1, Cmcon_SPEC, RW>

Compare Register Size for CMP1 MSIZE1. This bit field determines the number of bits in register CMP1 starting from bit 0 that are used for the compare operation with the System Timer. …

pub fn mstart1(self) -> RegisterField<24, 0x1f, 1, 0, Mstart1, Cmcon_SPEC, RW>

Start Bit Location for CMP1 MSTART1. This bit field determines the lowest bit number of the 64 bit STM that is compared with the content of register CMP1 bit 0. The number of bits to be compared is defined by bit field MSIZE1. …